% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{agg_ohlc}
\alias{agg_ohlc}
\title{Aggregate a time series of data into a single bar of \emph{OHLC} data.}
\usage{
agg_ohlc(t_series)
}
\arguments{
\item{\code{t_series}}{A \emph{matrix} or \emph{time series} with multiple
columns of data.}
}
\value{
A \emph{matrix} containing a single row, with the \emph{open},
  \emph{high}, \emph{low}, and \emph{close} values, and also the total
  \emph{volume} (if provided as either the second or fifth column of
  \code{t_series}).
}
\description{
Aggregate a time series of data into a single bar of \emph{OHLC} data.
}
\details{
The function \code{agg_ohlc()} aggregates a time series of data
  into a single bar of \emph{OHLC} data.
  It can accept either a single column of data or four columns of
  \emph{OHLC} data.
  It can also accept an additional column containing the trading volume.
  
The function \code{agg_ohlc()} calculates the \emph{open} value as equal to
the \emph{open} value of the first row of \code{t_series}.
  The \emph{high} value as the maximum of the \emph{high} column of
  \code{t_series}.
  The \emph{low} value as the minimum of the \emph{low} column of
  \code{t_series}.
  The \emph{close} value as the \emph{close} of the last row of
  \code{t_series}.
  The \emph{volume} value as the sum of the \emph{volume} column of
  \code{t_series}.

  For a single column of data, the \emph{open}, \emph{high}, \emph{low}, and
  \emph{close} values are all the same.
}
\examples{
\dontrun{
# Define matrix of OHLC data
oh_lc <- coredata(rutils::etf_env$VTI[, 1:5])
# Aggregate to single row matrix
ohlc_agg <- HighFreq::agg_ohlc(oh_lc)
# Compare with calculation in R
all.equal(drop(ohlc_agg),
  c(oh_lc[1, 1], max(oh_lc[, 2]), min(oh_lc[, 3]), oh_lc[NROW(oh_lc), 4], sum(oh_lc[, 5])), 
  check.attributes=FALSE)
}

}
