% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{agg_stats_r}
\alias{agg_stats_r}
\title{Calculate the aggregation (weighted average) of a statistical estimator over
a \emph{OHLC} time series using \code{R} code.}
\usage{
agg_stats_r(ohlc, calc_bars = "ohlc_variance", weighted = TRUE, ...)
}
\arguments{
\item{...}{additional parameters to the function \code{calc_bars}.}

\item{\code{ohlc}}{An \emph{OHLC} time series of prices and trading volumes,
in \emph{xts} format.}

\item{\code{calc_bars}}{A \emph{character} string representing a function
for calculating statistics for individual \emph{OHLC} bars.}

\item{\code{weighted}}{\emph{Boolean} argument: should estimate be weighted
by the trading volume? (default is \code{TRUE})}
}
\value{
A single \emph{numeric} value equal to the volume weighted average of
  an estimator over the time series.
}
\description{
Calculate the aggregation (weighted average) of a statistical estimator over
a \emph{OHLC} time series using \code{R} code.
}
\details{
The function \code{agg_stats_r()} calculates a single number
  representing the volume weighted average of statistics of individual
  \emph{OHLC} bars.
  It first calls the function \code{calc_bars} to calculate a vector of
  statistics for the \emph{OHLC} bars.
  For example, the statistic may simply be the difference between the
  \emph{High} minus \emph{Low} prices.  In this case the function
  \code{calc_bars} would calculate a vector of \emph{High} minus
  \emph{Low} prices.
  The function \code{agg_stats_r()} then calculates a trade volume
  weighted average of the vector of statistics.
  
  The function \code{agg_stats_r()} is implemented in \code{R} code.
}
\examples{
# Calculate weighted average variance for SPY (single number)
variance <- agg_stats_r(ohlc=HighFreq::SPY, calc_bars="ohlc_variance")
# Calculate time series of daily skew estimates for SPY
skew_daily <- apply.daily(x=HighFreq::SPY, FUN=agg_stats_r, calc_bars="ohlc_skew")

}
