% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{calc_cvar}
\alias{calc_cvar}
\title{Calculate the Value at Risk (\emph{VaR}) or the Conditional Value at Risk
(\emph{CVaR}) of an \emph{xts} \emph{time series} of returns, using \code{R}
code.}
\usage{
calc_cvar(tseries, method = "var", confi = pnorm(-2))
}
\arguments{
\item{\code{tseries}}{An \emph{xts} \emph{time series} of returns with
multiple columns.}

\item{\code{method}}{A \emph{string} specifying the type of risk measure
(the default is \code{method = "var"} - see Details).}

\item{\code{confi}}{The confidence level for calculating the
quantile (the default is \code{confi = pnorm(-2) = 0.02275}).}
}
\value{
A vector with the risk measures of the columns of the input
  \emph{time series} \code{tseries}.
}
\description{
Calculate the Value at Risk (\emph{VaR}) or the Conditional Value at Risk
(\emph{CVaR}) of an \emph{xts} \emph{time series} of returns, using \code{R}
code.
}
\details{
The function \code{calc_cvar()} calculates the Value at Risk (\emph{VaR})
  or the Conditional Value at Risk (\emph{CVaR}) of an \emph{xts} \emph{time
  series} of returns, using \code{R}
  
  The Value at Risk (\emph{VaR}) and the Conditional Value at Risk
  (\emph{CVaR}) are measures of the tail risk of returns.

  If \code{method = "var"} then \code{calc_cvar()} calculates the Value at
  Risk (\emph{VaR}) as the quantile of the returns as follows:
  \deqn{
    \alpha = \int_{-\infty}^{\mathrm{VaR}(\alpha)} \mathrm{f}(r) \, \mathrm{d}r
  }
  Where \eqn{\alpha} is the confidence level for calculating the quantile,
  and \eqn{\mathrm{f}(r)} is the probability density (distribution) of
  returns.
  
  If \code{method = "cvar"} then \code{calc_cvar()} calculates the Value at
  Risk (\emph{VaR}) as the Expected Tail Loss (\emph{ETL}) of the returns as
  follows:
  \deqn{
    \mathrm{CVaR} = \frac{1}{\alpha} \int_{0}^\alpha \mathrm{VaR}(p) \, \mathrm{d}p
  }
  Where \eqn{\alpha} is the confidence level for calculating the quantile.
}
\examples{
\dontrun{
# Calculate VTI and XLF returns
returns <- na.omit(rutils::etfenv$returns[, c("VTI", "XLF")])
# Calculate VaR
all.equal(HighFreq::calc_cvar(returns), 
  sapply(returns, quantile, probs=pnorm(-2)), check.attributes=FALSE)
# Calculate CVaR
all.equal(HighFreq::calc_cvar(returns, method="cvar", confi=0.02), 
  sapply(returns, function(x) mean(x[x < quantile(x, 0.02)])), 
  check.attributes=FALSE)
}

}
