% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_hurst}
\alias{calc_hurst}
\title{Calculate the Hurst exponent from the volatility ratio of aggregated returns.}
\usage{
calc_hurst(tseries, step = 1L)
}
\arguments{
\item{\code{tseries}}{A \emph{time series} or a \emph{matrix} of prices.}

\item{\code{step}}{The number of periods in each interval between
neighboring end points.}
}
\value{
The Hurst exponent calculated from the variance of aggregated
  returns.
}
\description{
Calculate the Hurst exponent from the volatility ratio of aggregated returns.
}
\details{
The function \code{calc_hurst()} calculates the Hurst exponent from the
  ratios of the volatilities of aggregated returns.

  The aggregated volatility \eqn{\sigma_t} scales (increases) with the
  length of the aggregation interval \eqn{\Delta t} raised to the power of
  the \emph{Hurst exponent} \eqn{H}:
    \deqn{
      \sigma_t = \sigma {\Delta t}^H
    }
  Where \eqn{\sigma} is the daily return volatility.

  The \emph{Hurst exponent} \eqn{H} is equal to the logarithm of the ratio
  of the volatilities divided by the logarithm of the time interval
  \eqn{\Delta t}:
    \deqn{
      H = \frac{\log{\sigma_t} - \log{\sigma}}{\log{\Delta t}}
    }

  The function \code{calc_hurst()} calls the function \code{calc_var_ag()}
  to calculate the aggregated volatility \eqn{\sigma_t}.
}
\examples{
\dontrun{
# Calculate the log prices
price_s <- na.omit(rutils::etf_env$price_s[, c("XLP", "VTI")])
price_s <- log(price_s)
# Calculate the Hurst exponent from 21 day aggregations
calc_hurst(price_s, step=21)
}

}
