% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_hurst_ohlc}
\alias{calc_hurst_ohlc}
\title{Calculate the Hurst exponent from the volatility ratio of aggregated
\emph{OHLC} prices.}
\usage{
calc_hurst_ohlc(
  ohlc,
  step,
  method = "yang_zhang",
  closel = 0L,
  scale = TRUE,
  index = 0L
)
}
\arguments{
\item{\code{ohlc}}{A \emph{time series} or a \emph{matrix} of \emph{OHLC}
prices.}

\item{\code{step}}{The number of time periods in each interval between
neighboring end points.}

\item{\code{method}}{A \emph{character string} representing the price range
estimator for calculating the variance.  The estimators include:
\itemize{
  \item "close" close-to-close estimator,
  \item "rogers_satchell" Rogers-Satchell estimator,
  \item "garman_klass" Garman-Klass estimator,
  \item "garman_klass_yz" Garman-Klass with account for close-to-open price jumps,
  \item "yang_zhang" Yang-Zhang estimator,
 }
 (The default is the \code{method = "yang_zhang"}.)}

\item{\code{closel}}{A \emph{vector} with the lagged \emph{close} prices
of the \emph{OHLC time series}.  This is an optional argument. (The
default is \code{closel = 0}).}

\item{\code{scale}}{\emph{Boolean} argument: Should the returns be divided
by the time index, the number of seconds in each period? (The default is
\code{scale = TRUE}).}

\item{\code{index}}{A \emph{vector} with the time index of the \emph{time
series}.  This is an optional argument (the default is \code{index = 0}).}
}
\value{
The Hurst exponent calculated from the volatility ratio of
  aggregated \emph{OHLC} prices.
}
\description{
Calculate the Hurst exponent from the volatility ratio of aggregated
\emph{OHLC} prices.
}
\details{
The function \code{calc_hurst_ohlc()} calculates the Hurst exponent from the
ratios of the variances of aggregated \emph{OHLC} prices.

  The aggregated volatility \eqn{\sigma_t} increases with the length of the
  aggregation interval \eqn{\Delta t}.
  The aggregated volatility increases as the length of the aggregation
  interval \eqn{\Delta t} raised to the power of the \emph{Hurst exponent}
  \eqn{H}:
    \deqn{
      \sigma_t = \sigma {\Delta t}^H
    }
  Where \eqn{\sigma} is the daily return volatility.

  The \emph{Hurst exponent} \eqn{H} is equal to the logarithm of the ratio
  of the volatilities divided by the logarithm of the time interval
  \eqn{\Delta t}:
    \deqn{
      H = \frac{\log{\sigma_t} - \log{\sigma}}{\log{\Delta t}}
    }

  The function \code{calc_hurst_ohlc()} calls the function
  \code{calc_var_ohlc_ag()} to calculate the aggregated variance
  \eqn{\sigma^2_t}.
}
\examples{
\dontrun{
# Calculate the log ohlc prices
ohlc <- log(rutils::etfenv$VTI)
# Calculate the Hurst exponent from 21 day aggregations
calc_hurst_ohlc(ohlc, step=21)
}  # end dontrun

}
