% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_kurtosis}
\alias{calc_kurtosis}
\title{Calculate the kurtosis of the columns of a \emph{time series} or a
\emph{matrix} using \code{RcppArmadillo}.}
\usage{
calc_kurtosis(se_ries, method = "kurt_mom", al_pha = 0.25)
}
\arguments{
\item{\code{se_ries}}{A \emph{time series} or a \emph{matrix} of data.}

\item{\code{method}}{A \emph{string} specifying the type of kurtosis (see
Details). (The default is the \code{method = "kurt_mom"}.)}

\item{\code{al_pha}}{The confidence level for calculating the quantiles.
(the default is \code{al_pha = 0.25}).}
}
\value{
A single-row matrix with the kurtosis of the columns of
  \code{se_ries}.
}
\description{
Calculate the kurtosis of the columns of a \emph{time series} or a
\emph{matrix} using \code{RcppArmadillo}.
}
\details{
The function \code{calc_kurtosis()} calculates the kurtosis of the columns of
  a \emph{time series} or a \emph{matrix} of data using \code{RcppArmadillo}
  \code{C++} code.

  If \code{method = "kurt_mom"} (the default) then \code{calc_kurtosis()}
  calculates the kurt_mom kurtosis using the third moment of the data.

  If \code{method = "kurt_quant"} then it calculates the kurtosis using the
  differences between the quantiles of the data.

  If \code{method = "kurt_np"} then it calculates the kurtosis as the
  difference between the mean of the data minus its median, divided by the
  standard deviation.
  
  If the number of rows of \code{se_ries} is less than \code{3} then it
  returns zeros.
  
  The code examples below compare the function \code{calc_kurtosis()} with the
  kurtosis calculated using \code{R} code.
}
\examples{
\dontrun{
# Calculate VTI returns
re_turns <- na.omit(rutils::etf_env$re_turns$VTI)
# Calculate the kurt_mom kurtosis
HighFreq::calc_kurtosis(re_turns)
# Compare HighFreq::calc_kurtosis() with kurt_mom kurtosis
calc_kurtosisr <- function(x) {
  x <- (x-mean(x)); nr <- NROW(x);
  nr*sum(x^3)/(var(x))^1.5/(nr-1)/(nr-2)
}  # end calc_kurtosisr
all.equal(HighFreq::calc_kurtosis(re_turns), 
  calc_kurtosisr(re_turns), check.attributes=FALSE)
# Compare the speed of RcppArmadillo with R code
library(microbenchmark)
summary(microbenchmark(
  Rcpp=HighFreq::calc_kurtosis(re_turns),
  Rcode=calc_kurtosisr(re_turns),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
# Calculate the kurt_quant kurtosis
HighFreq::calc_kurtosis(re_turns, method = "quantile", al_pha = 0.1)
# Compare HighFreq::calc_kurtosis() with quantile kurtosis
calc_kurtosisq <- function(x) {
  	quantile_s <- quantile(x, c(0.25, 0.5, 0.75), type=5)
  	(quantile_s[3] + quantile_s[1] - 2*quantile_s[2])/(quantile_s[3] - quantile_s[1])
}  # end calc_kurtosisq
all.equal(drop(HighFreq::calc_kurtosis(re_turns, method = "kurt_quant")), 
  calc_kurtosisq(re_turns), check.attributes=FALSE)
# Compare the speed of RcppArmadillo with R code
summary(microbenchmark(
  Rcpp=HighFreq::calc_kurtosis(re_turns, method = "kurt_quant"),
  Rcode=calc_kurtosisq(re_turns),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
# Calculate the kurt_np kurtosis
HighFreq::calc_kurtosis(re_turns, method = "kurt_np")
# Compare HighFreq::calc_kurtosis() with R nonparametric kurtosis
all.equal(drop(HighFreq::calc_kurtosis(re_turns, method = "kurt_np")), 
  (mean(re_turns)-median(re_turns))/sd(re_turns), 
  check.attributes=FALSE)
# Compare the speed of RcppArmadillo with R code
summary(microbenchmark(
  Rcpp=HighFreq::calc_kurtosis(re_turns, method = "kurt_np"),
  Rcode=(mean(re_turns)-median(re_turns))/sd(re_turns),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
