% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_lm}
\alias{calc_lm}
\title{Perform multivariate linear regression using least squares and return a
named list of regression coefficients, their t-values, and p-values.}
\usage{
calc_lm(respv, predm)
}
\arguments{
\item{\code{respv}}{A single-column \emph{time series} or a \emph{vector}
of response data.}

\item{\code{predm}}{A \emph{time series} or a \emph{matrix} of predictor
data.}
}
\value{
A named list with three elements: a \emph{matrix} of coefficients
  (named \emph{"coefficients"}), the \emph{z-score} of the last residual
  (named \emph{"zscore"}), and a \emph{vector} with the R-squared and
  F-statistic (named \emph{"stats"}). The numeric \emph{matrix} of
  coefficients named \emph{"coefficients"} contains the alpha and beta
  coefficients, and their \emph{t-values} and \emph{p-values}.
}
\description{
Perform multivariate linear regression using least squares and return a
named list of regression coefficients, their t-values, and p-values.
}
\details{
The function \code{calc_lm()} performs the same calculations as the
  function \code{lm()} from package \emph{stats}. 
  It uses \code{RcppArmadillo} \code{C++} code so it's several times faster
  than \code{lm()}. The code was inspired by this article (but it's not
  identical to it):
  http://gallery.rcpp.org/articles/fast-linear-model-with-armadillo/
}
\examples{
\dontrun{
# Calculate historical returns
retp <- na.omit(rutils::etfenv$returns[, c("XLF", "VTI", "IEF")])
# Response equals XLF returns
respv <- retp[, 1]
# Predictor matrix equals VTI and IEF returns
predm <- retp[, -1]
# Perform multivariate regression using lm()
regmod <- lm(respv ~ predm)
regsum <- summary(regmod)
# Add unit intercept column to the predictor matrix
predm <- cbind(rep(1, NROW(predm)), predm)
# Perform multivariate regression using calc_lm()
regarma <- HighFreq::calc_lm(respv=respv, predm=predm)
# Compare the outputs of both functions
all.equal(regarma$coefficients[, "coeff"], unname(coef(regmod)))
all.equal(unname(regarma$coefficients), unname(regsum$coefficients))
all.equal(unname(regarma$stats), c(regsum$r.squared, unname(regsum$fstatistic[1])))
# Compare the speed of RcppArmadillo with R code
summary(microbenchmark(
  Rcpp=HighFreq::calc_lm(respv=respv, predm=predm),
  Rcode=lm(respv ~ predm),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}  # end dontrun

}
