% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_ranks_stl}
\alias{calc_ranks_stl}
\title{Calculate the ranks of the elements of a single-column \emph{time series},
\emph{matrix}, or a \emph{vector} using \code{RcppArmadillo}.}
\usage{
calc_ranks_stl(timeser)
}
\arguments{
\item{\code{timeser}}{A single-column \emph{time series}, \emph{matrix}, or
a \emph{vector}.}
}
\value{
An \emph{integer vector} with the ranks of the elements of
  \code{timeser}.
}
\description{
Calculate the ranks of the elements of a single-column \emph{time series},
\emph{matrix}, or a \emph{vector} using \code{RcppArmadillo}.
}
\details{
The function \code{calc_ranks_stl()} calculates the ranks of the elements
  of a single-column \emph{time series}, \emph{matrix}, or a \emph{vector}.
  The function \code{calc_ranks_stl()} is slightly faster than the function
  \code{calc_ranks()}.
  
  The permutation index is an integer vector which sorts a given vector into
  ascending order. 
  The permutation index of the permutation index is the \emph{reverse}
  permutation index, because it sorts the vector from ascending order back
  into its original unsorted order.
  The ranks of the elements are equal to the \emph{reverse} permutation
  index. The function \code{calc_ranks()} calculates the \emph{reverse}
  permutation index.
  
  The ranks produced by \code{calc_ranks_stl()} start at zero, following the
  \code{C++} convention.

  The \code{STL} \code{C++} function \code{std::sort()} sorts a vector into
  ascending order. It can also be used to calculate the permutation index
  which sorts the vector into an ascending order.
  
  The function \code{calc_ranks_stl()} calls the function \code{std::sort()}
  twice:
  First, it calculates the permutation index which sorts the vector
  \code{timeser} into ascending order.
  Second, it calculates the permutation index of the permutation index,
  which are the ranks (the \emph{reverse} permutation index) of the vector
  \code{timeser}.
}
\examples{
\dontrun{
# Create a vector of data
datav <- rnorm(1e3)
# Calculate the ranks of the elements using R code and RcppArmadillo
all.equal(rank(datav), drop(HighFreq::calc_ranks_stl(datav))+1)
# Compare the speed of R code with RcppArmadillo
library(microbenchmark)
summary(microbenchmark(
  Rcode=rank(datav),
  Rcpp=HighFreq::calc_ranks_stl(datav),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}  # end dontrun

}
