% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{calc_rets}
\alias{calc_rets}
\title{Calculate returns of either \code{TAQ} or \code{OHLC} data in \code{xts} format.}
\usage{
calc_rets(x_ts)
}
\arguments{
\item{x_ts}{\code{xts} time series of \code{TAQ} or \code{OHLC} data.}
}
\value{
\code{xts} time series of returns and volumes.
}
\description{
Calculate returns of either \code{TAQ} or \code{OHLC} data in \code{xts} format.
}
\details{
The function \code{calc_rets} calculates returns and binds them with
  volume data.
}
\examples{
# create time index of one second intervals for a single day
in_dex <- seq(from=as.POSIXct("2015-02-09 09:30:00"), to=as.POSIXct("2015-02-09 16:00:00"), by="1 sec")
# create xts of random TAQ prices
x_ts <- xts(cumsum(rnorm(length(in_dex))), order.by=in_dex)
# create vector of random bid-offer prices
bid_offer <- abs(rnorm(length(in_dex)))/10
# create TAQ data using cbind
x_ts <- cbind(x_ts-bid_offer, x_ts+bid_offer)
# add Trade.Price
x_ts <- cbind(x_ts, x_ts+rnorm(length(in_dex))/10)
# add Volume
x_ts <- cbind(x_ts, sample(x=10*(2:18), size=length(in_dex), replace=TRUE))
colnames(x_ts) <- c("Bid.Price", "Ask.Price", "Trade.Price", "Volume")
x_ts <- calc_rets(x_ts)
}

