% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{calc_season}
\alias{calc_season}
\title{Calculate the intraday seasonality aggregations over a single-column
\emph{xts} time series.}
\usage{
calc_season(xtsv, endp = format(zoo::index(xtsv), "\%H:\%M"))
}
\arguments{
\item{\code{xtsv}}{A single-column \emph{xts} time series.}

\item{\code{endp}}{A vector of \emph{character} strings representing points in
time, of the same length as the argument \code{xtsv}.}
}
\value{
An \emph{xts} time series with mean aggregations over the seasonality
  interval.
}
\description{
Calculate the intraday seasonality aggregations over a single-column
\emph{xts} time series.
}
\details{
The function \code{calc_season()} calculates the mean of values
  observed at the same points in time specified by the argument
  \code{endp}. 
  An example of a daily seasonality aggregation is the average price of a
  stock between 9:30AM and 10:00AM every day, over many days. The argument
  \code{endp} is passed into function \code{tapply()}, and must be the same
  length as the argument \code{xtsv}.
}
\examples{
# Calculate running variance of each minutely OHLC bar of data
xtsv <- ohlc_variance(HighFreq::SPY)
# Remove overnight variance spikes at "09:31"
endp <- format(index(xtsv), "\%H:\%M")
xtsv <- xtsv[!endp=="09:31", ]
# Calculate daily seasonality of variance
var_seasonal <- calc_season(xtsv=xtsv)
chart_Series(x=var_seasonal, name=paste(colnames(var_seasonal),
  "daily seasonality of variance"))
  
}
