% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_var}
\alias{calc_var}
\title{Calculate the dispersion (variance) of the columns of a \emph{time series}
or a \emph{matrix} using \code{RcppArmadillo}.}
\usage{
calc_var(tseries, method = "moment")
}
\arguments{
\item{\code{tseries}}{A \emph{time series} or a \emph{matrix} of data.}

\item{\code{method}}{A \emph{string} specifying the type of dispersion model
(see Details). (The default is the \code{method = "moment"}.)}
}
\value{
A row vector equal to the dispersion of the columns of the matrix
  \code{tseries}.
}
\description{
Calculate the dispersion (variance) of the columns of a \emph{time series}
or a \emph{matrix} using \code{RcppArmadillo}.
}
\details{
The dispersion is a measure of the variability of the data.  Examples of
  dispersion are the variance and the Median Absolute Deviation (\emph{MAD}).

  The function \code{calc_var()} calculates the dispersion of the
  columns of a \emph{time series} or a \emph{matrix} of data using
  \code{RcppArmadillo} \code{C++} code.
  
  If \code{method = "moment"} (the default) then \code{calc_var()}
  calculates the dispersion as the second moment of the data \eqn{\sigma^2}
  (the variance).

  If \code{method = "moment"} then \code{calc_var()} performs the same
  calculation as the function \code{colVars()} from package
  \href{https://cran.r-project.org/web/packages/matrixStats/index.html}{matrixStats},
  but it's much faster because it uses \code{RcppArmadillo} \code{C++} code.

  If \code{method = "quantile"} then it calculates the dispersion as the
  Median Absolute Deviation (\emph{MAD}):
  \deqn{
    MAD = median(abs(x - median(x)))
  }
  It also multiplies the \emph{MAD} by a factor of \code{1.4826}, to make it
  comparable to the standard deviation.

  If \code{method = "quantile"} then \code{calc_var()} performs the same
  calculation as the function \code{stats::mad()}, but it's much faster
  because it uses \code{RcppArmadillo} \code{C++} code.

  If the number of rows of \code{tseries} is less than \code{3} then it
  returns zeros.
}
\examples{
\dontrun{
# Calculate VTI and XLF returns
re_turns <- na.omit(rutils::etf_env$re_turns[, c("VTI", "XLF")])
# Compare HighFreq::calc_var() with standard var()
all.equal(drop(HighFreq::calc_var(re_turns)), 
  apply(re_turns, 2, var), check.attributes=FALSE)
# Compare HighFreq::calc_var() with matrixStats
all.equal(drop(HighFreq::calc_var(re_turns)), 
  matrixStats::colVars(re_turns), check.attributes=FALSE)
# Compare the speed of RcppArmadillo with matrixStats and with R code
library(microbenchmark)
summary(microbenchmark(
  Rcpp=HighFreq::calc_var(re_turns),
  matrixStats=matrixStats::colVars(re_turns),
  Rcode=apply(re_turns, 2, var),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
# Compare HighFreq::calc_var() with stats::mad()
all.equal(drop(HighFreq::calc_var(re_turns, method="quantile")), 
  sapply(re_turns, mad), check.attributes=FALSE)
# Compare the speed of RcppArmadillo with stats::mad()
summary(microbenchmark(
  Rcpp=HighFreq::calc_var(re_turns, method="quantile"),
  Rcode=sapply(re_turns, mad),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
