% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_var_ohlc_ag}
\alias{calc_var_ohlc_ag}
\title{Calculate the variance of aggregated \emph{OHLC} prices using different
price range estimators.}
\usage{
calc_var_ohlc_ag(
  ohlc,
  step,
  method = "yang_zhang",
  closel = 0L,
  scale = TRUE,
  index = 0L
)
}
\arguments{
\item{\code{ohlc}}{A \emph{time series} or a \emph{matrix} of \emph{OHLC}
prices.}

\item{\code{step}}{The number of time periods in each interval between
neighboring end points.}

\item{\code{method}}{A \emph{character string} representing the price range
estimator for calculating the variance.  The estimators include:
\itemize{
  \item "close" close-to-close estimator,
  \item "rogers_satchell" Rogers-Satchell estimator,
  \item "garman_klass" Garman-Klass estimator,
  \item "garman_klass_yz" Garman-Klass with account for close-to-open price jumps,
  \item "yang_zhang" Yang-Zhang estimator,
 }
 (The default is the \code{method = "yang_zhang"}.)}

\item{\code{closel}}{A \emph{vector} with the lagged \emph{close} prices
of the \emph{OHLC time series}.  This is an optional argument. (The
default is \code{closel = 0}).}

\item{\code{scale}}{\emph{Boolean} argument: Should the returns be divided
by the time index, the number of seconds in each period? (The default is
\code{scale = TRUE}).}

\item{\code{index}}{A \emph{vector} with the time index of the \emph{time
series}.  This is an optional argument (the default is \code{index = 0}).}
}
\value{
The variance of aggregated \emph{OHLC} prices.
}
\description{
Calculate the variance of aggregated \emph{OHLC} prices using different
price range estimators.
}
\details{
The function \code{calc_var_ohlc_ag()} calculates the variance of
  \emph{OHLC} prices aggregated over the end points.

  It first calculates the end points spaced apart by the number of periods
  equal to the argument \code{step}.  Then it aggregates the \emph{OHLC}
  prices to the end points. Finally it calculates the variance of the
  aggregated \emph{OHLC} prices.

  The choice of the first end point is arbitrary, so \code{calc_var_ohlc_ag()}
  calculates the different end points for all the possible starting points.
  It then calculates the variance values for all the different end points
  and averages them.

  The aggregated volatility \eqn{\sigma_t} increases with the length of the
  aggregation interval \eqn{\Delta t}.
  The aggregated volatility increases as the length of the aggregation
  interval \eqn{\Delta t} raised to the power of the \emph{Hurst exponent}
  \eqn{H}:
    \deqn{
      \sigma_t = \sigma {\Delta t}^H
    }
  Where \eqn{\sigma} is the daily return volatility.

  The function \code{calc_var_ohlc_ag()} can therefore be used to calculate
  the \emph{Hurst exponent} from the variance ratio.
}
\examples{
\dontrun{
# Calculate the log ohlc prices
ohlc <- log(rutils::etfenv$VTI)
# Calculate the daily variance of percentage returns
calc_var_ohlc_ag(ohlc, step=1)
# Calculate the variance of returns aggregated over 21 days
calc_var_ohlc_ag(ohlc, step=21)
# The variance over 21 days is approximately 21 times the daily variance
21*calc_var_ohlc_ag(ohlc, step=1)
}  # end dontrun

}
