% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{calc_var_ohlc_r}
\alias{calc_var_ohlc_r}
\title{Calculate the variance of an \emph{OHLC} time series, using different range
estimators for variance.}
\usage{
calc_var_ohlc_r(oh_lc, calc_method = "yang_zhang", scal_e = TRUE)
}
\arguments{
\item{\code{oh_lc}}{An \emph{OHLC} time series of prices in \emph{xts} format.}

\item{\code{calc_method}}{A \emph{character} string representing the method for
estimating variance.  The methods include:
\itemize{
  \item "close" close to close,
  \item "garman_klass" Garman-Klass,
  \item "garman_klass_yz" Garman-Klass with account for close-to-open price jumps,
  \item "rogers_satchell" Rogers-Satchell,
  \item "yang_zhang" Yang-Zhang,
 }
 (default is \code{"yang_zhang"})}

\item{\code{scal_e}}{\emph{Boolean} argument: should the returns be divided by the
number of seconds in each period? (default is \code{TRUE})}
}
\value{
A single \emph{numeric} value equal to the variance.
}
\description{
Calculate the variance of an \emph{OHLC} time series, using different range
estimators for variance.
}
\details{
The function \code{calc_var_ohlc_r()} calculates the variance
  from all the different intra-day and day-over-day returns (defined as the
  differences of \emph{OHLC} prices), using several different variance
  estimation methods.

  The default method is \code{"yang_zhang"}, which theoretically has the
  lowest standard error among unbiased estimators.
  The methods \code{"close"}, \code{"garman_klass_yz"}, and
  \code{"yang_zhang"} do account for close-to-open price jumps, while the
  methods \code{"garman_klass"} and \code{"rogers_satchell"} do not account
  for close-to-open price jumps.

  If \code{scal_e} is \code{TRUE} (the default), then the returns are divided
  by the differences of the time index (which scales the variance to the
  units of variance per second squared.) This is useful when calculating the
  variance from minutely bar data, because dividing returns by the number of
  seconds decreases the effect of overnight price jumps. If the time index is
  in days, then the variance is equal to the variance per day squared.
  
  The function \code{calc_var_ohlc_r()} is implemented in \code{R} code.
}
\examples{
# Calculate the variance of SPY returns
HighFreq::calc_var_ohlc_r(HighFreq::SPY, calc_method="yang_zhang")
# Calculate variance without accounting for overnight jumps
HighFreq::calc_var_ohlc_r(HighFreq::SPY, calc_method="rogers_satchell")
# Calculate the variance without scaling the returns
HighFreq::calc_var_ohlc_r(HighFreq::SPY, scal_e=FALSE)
}
