% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_var_vec}
\alias{calc_var_vec}
\title{Calculate the variance of a a single-column \emph{time series} or a
\emph{vector} using \code{RcppArmadillo}.}
\usage{
calc_var_vec(se_ries)
}
\arguments{
\item{\code{se_ries}}{A single-column \emph{time series} or a \emph{vector}.}
}
\value{
A \emph{numeric} value equal to the variance of the \emph{vector}.
}
\description{
Calculate the variance of a a single-column \emph{time series} or a
\emph{vector} using \code{RcppArmadillo}.
}
\details{
The function \code{calc_var_vec()} calculates the variance of a
  \emph{vector} using \code{RcppArmadillo} \code{C++} code, so it's
  significantly faster than the \code{R} function \code{var()}.
}
\examples{
\dontrun{
# Create a vector of random returns
re_turns <- rnorm(1e6)
# Compare calc_var_vec() with standard var()
all.equal(HighFreq::calc_var_vec(re_turns), 
  var(re_turns))
# Compare the speed of RcppArmadillo with R code
library(microbenchmark)
summary(microbenchmark(
  Rcpp=HighFreq::calc_var_vec(re_turns),
  Rcode=var(re_turns),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
