% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_varvec}
\alias{calc_varvec}
\title{Calculate the variance of a single-column \emph{time series} or a
\emph{vector} using \code{RcppArmadillo}.}
\usage{
calc_varvec(tseries)
}
\arguments{
\item{\code{tseries}}{A single-column \emph{time series} or a \emph{vector}.}
}
\value{
A \emph{numeric} value equal to the variance of the \emph{vector}.
}
\description{
Calculate the variance of a single-column \emph{time series} or a
\emph{vector} using \code{RcppArmadillo}.
}
\details{
The function \code{calc_varvec()} calculates the variance of a
  \emph{vector} using \code{RcppArmadillo} \code{C++} code, so it's
  significantly faster than the \code{R} function \code{var()}.
}
\examples{
\dontrun{
# Create a vector of random returns
returns <- rnorm(1e6)
# Compare calc_varvec() with standard var()
all.equal(HighFreq::calc_varvec(returns), 
  var(returns))
# Compare the speed of RcppArmadillo with R code
library(microbenchmark)
summary(microbenchmark(
  Rcpp=HighFreq::calc_varvec(returns),
  Rcode=var(returns),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
