% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diff_vec}
\alias{diff_vec}
\title{Calculate the differences between the neighboring elements of a
single-column \emph{time series} or a \emph{vector}.}
\usage{
diff_vec(se_ries, lagg = 1L, padd = TRUE)
}
\arguments{
\item{\code{se_ries}}{A single-column \emph{time series} or a \emph{vector}.}

\item{\code{lagg}}{An \emph{integer} equal to the number of time periods to
lag when calculating the differences (the default is \code{lagg = 1}).}

\item{\code{padd}}{\emph{Boolean} argument: Should the output \emph{vector}
be padded (extended) with zeros, in order to return a \emph{vector} of the
same length as the input? (the default is \code{padd = TRUE})}
}
\value{
A column \emph{vector} containing the differences between the
  elements of the input vector.
}
\description{
Calculate the differences between the neighboring elements of a
single-column \emph{time series} or a \emph{vector}.
}
\details{
The function \code{diff_vec()} calculates the differences between the
  input \emph{time series} or \emph{vector} and its lagged version.
  
  The argument \code{lagg} specifies the number of lags.  For example, if
  \code{lagg=3} then the differences will be taken between each element
  minus the element three time periods before it (in the past).  The default
  is \code{lagg = 1}.

  The argument \code{padd} specifies whether the output \emph{vector} should
  be padded (extended) with zeros at the beginning, in order to return a
  \emph{vector} of the same length as the input.  The default is
  \code{padd = TRUE}. The padding operation can be time-consuming, because it
  requires the copying of data.
  
  The function \code{diff_vec()} is implemented in \code{RcppArmadillo}
  \code{C++} code, which makes it several times faster than \code{R} code.
}
\examples{
\dontrun{
# Create a vector of random returns
re_turns <- rnorm(1e6)
# Compare diff_vec() with rutils::diff_it()
all.equal(drop(HighFreq::diff_vec(re_turns, lagg=3, padd=TRUE)),
  rutils::diff_it(re_turns, lagg=3))
# Compare the speed of RcppArmadillo with R code
library(microbenchmark)
summary(microbenchmark(
  Rcpp=HighFreq::diff_vec(re_turns, lagg=3, padd=TRUE),
  Rcode=rutils::diff_it(re_turns, lagg=3),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
