% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diffit}
\alias{diffit}
\title{Calculate the row differences of a \emph{time series} or a \emph{matrix}
using \emph{RcppArmadillo}.}
\usage{
diffit(tseries, lagg = 1L, pad_zeros = TRUE)
}
\arguments{
\item{\code{tseries}}{A \emph{time series} or a \emph{matrix}.}

\item{\code{lagg}}{An \emph{integer} equal to the number of rows (time
periods) to lag when calculating the differences (the default is
\code{lagg = 1}).}

\item{\code{pad_zeros}}{\emph{Boolean} argument: Should the output
\emph{matrix} be padded (extended) with zeros, in order to return a
\emph{matrix} with the same number of rows as the input? (the default is
\code{pad_zeros = TRUE})}
}
\value{
A \emph{matrix} containing the differences between the rows of the
  input \emph{matrix} \code{tseries}.
}
\description{
Calculate the row differences of a \emph{time series} or a \emph{matrix}
using \emph{RcppArmadillo}.
}
\details{
The function \code{diffit()} calculates the differences between the rows
  of the input \emph{matrix} \code{tseries} and its lagged version.
  
  The argument \code{lagg} specifies the number of lags applied to the rows
  of the lagged version of \code{tseries}. 
  For positive \code{lagg} values, the lagged version of \code{tseries} has
  its rows shifted \emph{forward} (down) by the number equal to \code{lagg}
  rows. For negative \code{lagg} values, the lagged version of
  \code{tseries} has its rows shifted \emph{backward} (up) by the number
  equal to \code{-lagg} rows.
  For example, if \code{lagg=3} then the lagged version will have its rows
  shifted down by \code{3} rows, and the differences will be taken between
  each row minus the row three time periods before it (in the past). The
  default is \code{lagg = 1}.

  The argument \code{pad_zeros} specifies whether the output \emph{matrix}
  should be padded (extended) with the rows of the initial (warmup) period
  at the front, in order to return a \emph{matrix} with the same number of
  rows as the input \code{tseries}.  The default is \code{pad_zeros = TRUE}.
  The padding operation can be time-consuming, because it requires the
  copying of data.
  
  The function \code{diffit()} is implemented in \code{RcppArmadillo}
  \code{C++} code, which makes it much faster than \code{R} code.
}
\examples{
\dontrun{
# Create a matrix of random data
datav <- matrix(sample(15), nc=3)
# Calculate differences with lagged rows
HighFreq::diffit(datav, lagg=2)
# Calculate differences with advanced rows
HighFreq::diffit(datav, lagg=-2)
# Compare HighFreq::diffit() with rutils::diffit()
all.equal(HighFreq::diffit(datav, lagg=2), 
  rutils::diffit(datav, lagg=2), 
  check.attributes=FALSE)
# Compare the speed of RcppArmadillo with R code
library(microbenchmark)
summary(microbenchmark(
  Rcpp=HighFreq::diffit(datav, lagg=2),
  Rcode=rutils::diffit(datav, lagg=2),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
