% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{do_call_rbind}
\alias{do_call_rbind}
\title{Recursively \sQuote{\code{rbind}} a list of objects, such as \code{xts} time series.}
\usage{
do_call_rbind(list_var)
}
\arguments{
\item{list_var}{list of \code{vectors}, \code{matrices}, \code{data
frames}, or \code{time series}.}
}
\value{
single \code{vector}, \code{matrix}, \code{data frame}, or
  \code{time series}.
}
\description{
Performs the same operation as \code{do.call(rbind, list_var)}, but using
recursion, which is much faster and uses less memory. This is the same
function as \sQuote{\code{\link[qmao]{do.call.rbind}}} from package
\sQuote{\href{https://r-forge.r-project.org/R/?group_id=1113}{qmao}}.
}
\details{
Calls lapply in a loop, each time binding neighboring elements and
  dividing the length of \code{list_var} by half. The result of performing
  \code{do_call_rbind(list_xts)} on a list of \code{xts} time series is
  identical to performing \code{do.call(rbind, list_xts)}. But
  \code{do.call(rbind, list_xts)} is very slow, and often causes an \sQuote{out of
  memory} error.
}
\examples{
# create xts time series
x_ts <- xts(x=rnorm(1000), order.by=(Sys.time()-3600*(1:1000)))
# split time series into daily list
list_xts <- split(x_ts, "days")
# rbind the list back into a time series and compare with the original
identical(x_ts, do_call_rbind(list_xts))
}

