% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{extreme_values}
\alias{extreme_values}
\title{Identify extreme values in a univariate \code{xts} time series.}
\usage{
extreme_values(time_series, vol_window = 51, vol_mult = 2)
}
\arguments{
\item{time_series}{univariate \code{xts} time series.}

\item{vol_window}{number of data points for estimating running volatility.}

\item{vol_mult}{volatility multiplier.}
}
\value{
\code{logical vector}.
}
\description{
Identifies extreme values as those that exceed a multiple of the running volatility.
}
\details{
Calculates running volatility as a quantile of values over a sliding
  window. Extreme values are those that exceed the product of the volatility
  multiplier times the running volatility. Extreme values are the very tips
  of the tails when the distribution of values becomes very fat-tailed. The
  volatility multiplier \code{vol_mult} controls the threshold at which
  values are identified as extreme. Smaller volatility multiplier values will
  cause more values to be identified as extreme.
}
\examples{
# create xts time series
x_ts <- xts(x=rnorm(1000), order.by=(Sys.time()-3600*(1:1000)))
# scrub extreme values
x_ts <- x_ts[!extreme_values(x_ts, vol_mult=1)]
}

