% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{get_symbols}
\alias{get_symbols}
\title{Download time series data from an external source (by default \code{OHLC}
prices from \code{YAHOO}), and save it into an environment.}
\usage{
get_symbols(sym_bols, env_out, start_date = "2007-01-01",
  end_date = Sys.Date())
}
\arguments{
\item{sym_bols}{vector of strings representing instrument symbols (tickers).}

\item{env_out}{environment for saving the data.}

\item{start_date}{start date of time series data.  (default is "2007-01-01")}

\item{end_date}{end date of time series data.  (default is \code{Sys.Date()})}
}
\value{
A vector of \code{sym_bols} returned invisibly.
}
\description{
Download time series data from an external source (by default \code{OHLC}
prices from \code{YAHOO}), and save it into an environment.
}
\details{
The function \code{get_symbols} downloads \code{OHLC} prices from 
  \code{YAHOO} into an environment, adjusts the prices, and saves them back 
  to that environment. The function \code{get_symbols()} calls the function 
  \code{getSymbols.yahoo()} to download the \code{OHLC} prices, and performs 
  a similar operation to the function \code{getSymbols()} from package 
  \href{https://cran.r-project.org/web/packages/quantmod/index.html}{quantmod}.
   But \code{get_symbols()} is faster (because it's more specialized), and is
  able to handle symbols like "LOW", which \code{getSymbols()} can't handle
  because the function \code{Lo()} can't handle them. The \code{start_date}
  and \code{end_date} must be either of class \code{Date}, or a string in the
  format "YYYY-mm-dd".
  \code{get_symbols()} returns invisibly the vector of \code{sym_bols}.
}
\examples{
\dontrun{
new_env <- new.env()
get_symbols(sym_bols=c("MSFT", "XOM"),
            env_out=new_env, 
            start_date="2012-12-01",
            end_date="2015-12-01")
}
}

