\name{HighFreq-package}
\alias{HighFreq-package}
\alias{HighFreq}
\docType{package}
\title{
HighFreq R package
}
\description{
Read, write, chain, join, scrub, and aggregate high frequency TAQ and OHLC data.
Functions for chaining and joining time series, scrubbing bad data, managing time zones and alligning time indices, converting TAQ data to OHLC format, aggregating data to lower frequency.
}
\details{
\tabular{ll}{
Package: \tab HighFreq\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-02-08\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Jerzy Pawlowski
Maintainer: Jerzy Pawlowski <algoquant@algoquants.ch>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
