% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{hurst_ohlc}
\alias{hurst_ohlc}
\title{Calculate time series of Hurst exponent estimates for a \code{OHLC} time
series.}
\usage{
hurst_ohlc(ohlc, calc_method = "rogers.satchell")
}
\arguments{
\item{ohlc}{\code{OHLC} time series of prices.}

\item{calc_method}{\code{character} string representing method for estimating
Hurst exponent.}
}
\value{
Time series of Hurst exponent estimates.
}
\description{
Calculate time series of Hurst exponent estimates for a \code{OHLC} time
series.
}
\details{
Calculates Hurst exponent estimates from \code{OHLC} prices at each
  point in time (bar).  The methods include Garman-Klass and Rogers-Satchell.
}
\examples{
# create time index of one second intervals for a single day
in_dex <- seq(from=as.POSIXct("2015-02-09 09:30:00"),
              to=as.POSIXct("2015-02-09 16:00:00"), by="1 sec")
# create xts of random prices
x_ts <- xts(cumsum(rnorm(length(in_dex))), order.by=in_dex)
# aggregate to minutes OHLC data
oh_lc <- to.period(x=x_ts, period="minutes")
# calculate Hurst exponent
hurst_exp <- hurst_ohlc(oh_lc)
}

