% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_reg}
\alias{inv_reg}
\title{Calculate a time series of variance estimates over a rolling look-back interval
for an \emph{OHLC} time series of prices, using different range estimators
for variance.}
\usage{
inv_reg(re_turns, max_eigen)
}
\arguments{
\item{vec_tor}{A numeric \emph{vector} of data.}

\item{look_back}{The length of the look-back interval, equal to the number of rows
of data used for calculating the variance.}
}
\value{
A numeric \emph{vector} of the same length as the argument \code{vec_tor}.
}
\description{
Calculate a time series of variance estimates over a rolling look-back interval
for an \emph{OHLC} time series of prices, using different range estimators
for variance.
}
\details{
The function \code{inv_reg()} calculates a \emph{vector} of rolling 
  variance estimates, from over a \emph{vector} of returns, using \emph{Rcpp}.
}
\examples{
\dontrun{
# create minutely OHLC time series of random prices
vec_tor <- HighFreq::random_ohlc()
# calculate variance estimates for vec_tor over a 21 period interval
var_rolling <- HighFreq::inv_reg(vec_tor, look_back=21)
# calculate variance estimates for SPY
var_rolling <- HighFreq::inv_reg(HighFreq::SPY, calc_method="yang_zhang")
# calculate SPY variance without accounting for overnight jumps
var_rolling <- HighFreq::inv_reg(HighFreq::SPY, calc_method="rogers_satchell")
}
}
