% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{jump_values}
\alias{jump_values}
\title{Identify jump values in a univariate xts time series.}
\usage{
jump_values(time_series, vol_window = 51, thresh_old = 2)
}
\arguments{
\item{time_series}{univariate \code{xts} time series.}

\item{vol_window}{number of data points for estimating running volatility.}

\item{thresh_old}{level of change above which values are considered extreme.}
}
\value{
\code{logical vector}.
}
\description{
Calculates the running volatility of simple returns, defined as a quantile over a sliding window.
Identifies jump values as those returns that exceed the scaled running volatility, but their sum with the lagged returns doesn't exceed it.
}
\details{
Jump values are those where neighboring values both exceed the scaled running volatility (defined as a quantile), but their sum doesn't exceed it.
}
\examples{
# create xts time series
x_ts <- xts(x=rnorm(1000), order.by=(Sys.time()-3600*(1:1000)))
# scrub jump values
x_ts <- x_ts[!jump_values(x_ts, thresh_old=1.0)]
}

