% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{moment_ohlc}
\alias{moment_ohlc}
\title{Calculate the moment of a \code{OHLC} time series.}
\usage{
moment_ohlc(ohlc, mom_fun = "vol_ohlc", calc_method = "rogers.satchell",
  vo_lu = TRUE, ...)
}
\arguments{
\item{ohlc}{\code{OHLC} time series of prices.}

\item{mom_fun}{\code{character} string representing function for estimating
the moment.}

\item{calc_method}{\code{character} string representing method for estimating
the moment.}

\item{vo_lu}{\code{logical} should estimate be weighted by trade volume.}
}
\value{
\code{numeric} value with estimate of the moment.
}
\description{
Calculate the moment of a \code{OHLC} time series.
}
\details{
Calculates a single number representing the estimate of the moment
  of a \code{OHLC} time series of prices.  By default the estimate is trade
  volume weighted.
}
\examples{
# create time index of one minute intervals over several days
in_dex <- seq(from=as.POSIXct("2015-02-09 09:30:00"),
              to=as.POSIXct("2015-02-28 16:00:00"), by="1 min")
# create xts of random prices
x_ts <- xts(exp(cumsum(0.001*rnorm(length(in_dex)))), order.by=in_dex)
# add trade Volume data
x_ts <- merge(x_ts,
          volume=sample(x=10*(2:18),
            size=length(in_dex), replace=TRUE))
# aggregate to hours OHLC data
oh_lc <- to.period(x=x_ts, period="hours")
# calculate time series of daily skew estimates
sk_ew <- apply.daily(x=oh_lc, FUN=moment_ohlc, mom_fun="skew_ohlc")
}

