% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mult_vec_mat}
\alias{mult_vec_mat}
\title{Multiply in place (without copying) the columns or rows of a \emph{matrix}
times a \emph{vector}, element-wise.}
\usage{
mult_vec_mat(vector, matrix, by_col = TRUE)
}
\arguments{
\item{\code{vector}}{A \emph{vector}.}

\item{\code{matrix}}{A \emph{matrix}.}

\item{\code{by_col}}{A \emph{Boolean} argument: if \code{TRUE} then multiply
the columns, otherwise multiply the rows (the default is
\code{by_col = TRUE}.)}
}
\value{
A single \emph{integer} value, equal to either the number of
  \emph{matrix} columns or the number of rows.
}
\description{
Multiply in place (without copying) the columns or rows of a \emph{matrix}
times a \emph{vector}, element-wise.
}
\details{
The function \code{mult_vec_mat()} multiplies the columns or rows of a
  \emph{matrix} times a \emph{vector}, element-wise.

  If the number of \emph{vector} elements is equal to the number of matrix
  columns, then it multiplies the columns by the \emph{vector}, and returns
  the number of columns. If the number of \emph{vector} elements is equal to
  the number of rows, then it multiplies the rows, and returns the number of
  rows.

  If the \emph{matrix} is square and if \code{by_col} is \code{TRUE} then it
  multiplies the columns, otherwise it multiplies the rows.
  
  It accepts \emph{pointers} to the \emph{matrix} and \emph{vector}, and
  replaces the old \emph{matrix} values with the new values.
  It performs the calculation in place, without copying the \emph{matrix} in
  memory (which greatly increases the computation speed).
  It performs an implicit loop over the \emph{matrix} rows and columns using
  the \emph{Armadillo} operators \code{each_row()} and \code{each_col()},
  instead of performing explicit \code{for()} loops (both methods are
  equally fast).

  The function \code{mult_vec_mat()} uses \code{RcppArmadillo} \code{C++}
  code, so when multiplying large \emph{matrix} columns it's several times
  faster than vectorized \code{R} code, and it's even much faster compared
  to \code{R} when multiplying the \emph{matrix} rows.
}
\examples{
\dontrun{
# Multiply matrix columns using R
matrixv <- matrix(round(runif(25e4), 2), nc=5e2)
vectorv <- round(runif(5e2), 2)
prod_uct <- vectorv*matrixv
# Multiply the matrix in place
HighFreq::mult_vec_mat(vectorv, matrixv)
all.equal(prod_uct, matrixv)
# Compare the speed of Rcpp with R code
library(microbenchmark)
summary(microbenchmark(
    Rcpp=HighFreq::mult_vec_mat(vectorv, matrixv),
    Rcode=vectorv*matrixv,
    times=10))[, c(1, 4, 5)]  # end microbenchmark summary

# Multiply matrix rows using R
matrixv <- matrix(round(runif(25e4), 2), nc=5e2)
vectorv <- round(runif(5e2), 2)
prod_uct <- t(vectorv*t(matrixv))
# Multiply the matrix in place
HighFreq::mult_vec_mat(vectorv, matrixv, by_col=FALSE)
all.equal(prod_uct, matrixv)
# Compare the speed of Rcpp with R code
library(microbenchmark)
summary(microbenchmark(
    Rcpp=HighFreq::mult_vec_mat(vectorv, matrixv, by_col=FALSE),
    Rcode=t(vectorv*t(matrixv)),
    times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
