% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{ohlc_sharpe}
\alias{ohlc_sharpe}
\title{Calculate time series of point Sharpe-like statistics for each row of a
\emph{OHLC} time series.}
\usage{
ohlc_sharpe(ohlc, method = "close")
}
\arguments{
\item{\code{ohlc}}{An \emph{OHLC} time series of prices in \emph{xts} format.}

\item{\code{method}}{A \emph{character} string representing method for
estimating the Sharpe-like exponent.}
}
\value{
An \emph{xts} time series with the same number of rows as the
  argument \code{ohlc}.
}
\description{
Calculate time series of point Sharpe-like statistics for each row of a
\emph{OHLC} time series.
}
\details{
The function \code{ohlc_sharpe()} calculates Sharpe-like statistics
  for each row of a \emph{OHLC} time series.
  The Sharpe-like statistic is defined as the ratio of the difference between
  \emph{Close} minus \emph{Open} prices divided by the difference between
  \emph{High} minus \emph{Low} prices.
  This statistic may also be interpreted as something like a \emph{Hurst
  exponent} for a single row of data.
  The motivation for the Sharpe-like statistic is the notion that if prices
  are trending in the same direction inside a given time bar of data, then
  this statistic is close to either 1 or -1.
}
\examples{
# Calculate time series of running Sharpe ratios for SPY
sharpe_running <- ohlc_sharpe(HighFreq::SPY)
}
