% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{open_close}
\alias{open_close}
\title{Adjusts an \code{OHLC} time series to make open prices equal to the close
prices from the previous period.}
\usage{
open_close(x_ts)
}
\arguments{
\item{x_ts}{an \code{xts} time series containing one or more columns of data.}
}
\value{
\code{xts} \code{OHLC} time series with open prices equal to the
  close prices from the previous period.
}
\description{
Adjusts an \code{OHLC} time series to make open prices equal to the close
prices from the previous period.
}
\details{
Adds or subtracts a price adjustment to make all open prices equal
  to the close prices from the previous period.  The adjustment preserves the
  price differences within each bar of \code{OHLC} prices, and so preserves
  open to close returns, variance estimates, etc.
}
\examples{
# define end points at 10-minute intervals (SPY is minutely bars)
end_points <- rutils::end_points(SPY["2009"], inter_val=10)
# aggregate over 10-minute end_points:
open_close(x_ts=SPY["2009"], end_points=end_points)
# aggregate over days:
open_close(x_ts=SPY["2009"], period="days")
# equivalent to:
to.period(x=SPY["2009"], period="days", name=rutils::na_me(SPY))
}

