% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{push_cov2cor}
\alias{push_cov2cor}
\title{Calculate the correlation matrix from the covariance matrix.}
\usage{
push_cov2cor(covmat)
}
\arguments{
\item{\code{covmat}}{A \emph{matrix} of covariances.}
}
\value{
Void (no return value - modifies the covariance matrix in place).
}
\description{
Calculate the correlation matrix from the covariance matrix.
}
\details{
The function \code{push_cov2cor()} calculates the correlation matrix from
  the covariance matrix, in place, without copying the data in memory.
  
  The function \code{push_cov2cor()} accepts a \emph{pointer} to the
  covariance matrix, and it overwrites it with the correlation matrix.

  The function \code{push_cov2cor()} is written in \code{RcppArmadillo}
  \code{C++} so it's much faster than \code{R} code.
}
\examples{
\dontrun{
# Calculate a time series of returns
retp <- na.omit(rutils::etfenv$returns[, c("IEF", "VTI", "DBC")])
# Calculate the covariance matrix of returns
covmat <- cov(retp)
# Calculate the correlation matrix of returns
push_cov2cor(covmat)
all.equal(covmat, cor(retp))
}

}
