% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{random_ohlc}
\alias{random_ohlc}
\title{Simulate a random \emph{OHLC} time series of prices and trading volumes, in
\emph{xts} format.}
\usage{
random_ohlc(
  ohlc = NULL,
  reducit = TRUE,
  volat = 6.5e-05,
  drift = 0,
  datev = seq(from = as.POSIXct(paste(Sys.Date() - 3, "09:30:00")), to =
    as.POSIXct(paste(Sys.Date() - 1, "16:00:00")), by = "1 sec"),
  ...
)
}
\arguments{
\item{\code{ohlc}}{An \emph{OHLC} time series of prices and trading volumes,
in \emph{xts} format (default is \emph{NULL}).}

\item{\code{volat}}{The volatility per period of the \code{datev} time index
(default is \code{6.5e-05} per second, or about \code{0.01=1.0\%} per day).}

\item{\code{drift}}{The drift per period of the \code{datev} time index
(default is 0.0).}

\item{\code{datev}}{The time index for the \emph{OHLC} time series.}

\item{\code{reducit}}{\emph{Boolean} argument: should \code{ohlc} time series
be transformed to reduced form? (default is \code{TRUE})}
}
\value{
An \emph{xts} time series with the same dimensions and the same time
  index as the input \code{ohlc} time series.
}
\description{
Simulate a random \emph{OHLC} time series either by simulating random prices
following geometric Brownian motion, or by randomly sampling from an input
time series.
}
\details{
If the input \code{ohlc} time series is \emph{NULL} (the default),
  then the function \code{random_ohlc()} simulates a minutely \emph{OHLC}
  time series of random prices following geometric Brownian motion, over the
  two previous calendar days.

  If the input \code{ohlc} time series is not \emph{NULL}, then the rows of
  \code{ohlc} are randomly sampled, to produce a random time series.

  If \code{reducit} is \code{TRUE} (the default), then the \code{ohlc} time
  series is first transformed to reduced form, then randomly sampled, and
  finally converted to standard form.

  Note: randomly sampling from an intraday time series over multiple days
  will cause the overnight price jumps to be re-arranged into intraday price
  jumps.  This will cause moment estimates to become inflated compared to the
  original time series.
}
\examples{
# Create minutely synthetic OHLC time series of random prices
ohlc <- HighFreq::random_ohlc()
# Create random time series from SPY by randomly sampling it
ohlc <- HighFreq::random_ohlc(ohlc=HighFreq::SPY["2012-02-13/2012-02-15"])

}
