% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{random_taq}
\alias{random_taq}
\title{Calculate a random \emph{TAQ} time series of prices and trading volumes, in
\emph{xts} format.}
\usage{
random_taq(
  volat = 6.5e-05,
  drift = 0,
  indeks = seq(from = as.POSIXct(paste(Sys.Date() - 3, "09:30:00")), to =
    as.POSIXct(paste(Sys.Date() - 1, "16:00:00")), by = "1 sec"),
  bid_offer = 0.001,
  ...
)
}
\arguments{
\item{bid_offer}{The bid-offer spread expressed as a fraction of the prices
(default is 0.001=10bps).}

\item{\code{volat}}{The volatility per period of the \code{indeks} time index
(default is \code{6.5e-05} per second, or about \code{0.01=1.0\%} per day).}

\item{\code{drift}}{The drift per period of the \code{indeks} time index (default
is 0.0).}

\item{\code{indeks}}{The time index for the \emph{TAQ} time series.}
}
\value{
An \emph{xts} time series, with time index equal to the input
  \code{indeks} time index, and with four columns containing the bid, ask,
  and trade prices, and the trade volume.
}
\description{
Calculate a \emph{TAQ} time series of random prices following geometric
Brownian motion, combined with random trading volumes.
}
\details{
The function \code{random_taq()} calculates an \emph{xts} time
  series with four columns containing random prices following geometric
  Brownian motion: the bid, ask, and trade prices, combined with random trade
  volume data.
  If \code{indeks} isn't supplied as an argument, then by default it's
  equal to the secondly index over the two previous calendar days.
}
\examples{
# Create secondly TAQ time series of random prices
taq <- HighFreq::random_taq()
# Create random TAQ time series from SPY index
taq <- HighFreq::random_taq(indeks=index(HighFreq::SPY["2012-02-13/2012-02-15"]))

}
