% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{roll_agg}
\alias{roll_agg}
\title{Calculate the rolling aggregations of a statistical estimator over a
\code{OHLC} time series.}
\usage{
roll_agg(ohlc, esti_mator = "vari_ance", n = 20, N = 260,
  weight_ed = TRUE, ...)
}
\arguments{
\item{ohlc}{\code{OHLC} time series of prices and trading volumes.}

\item{esti_mator}{\code{character} string representing function for
estimating the moment.}

\item{n}{\code{numeric} number of periods for averaging of estimates.}

\item{N}{\code{numeric} number of periods in a year (to annualize the
estimates).}

\item{weight_ed}{\code{Boolean} should estimate be weighted by trade volume
(default \code{TRUE})?}

\item{...}{additional parameters to esti_mator function.}
}
\value{
\code{numeric} time series of rolling aggregations, with the same
  number of rows as the input \code{xts} series.
}
\description{
Calculate the rolling aggregations of a statistical estimator over a
\code{OHLC} time series.
}
\details{
Calculates a time series of rolling aggregations of an estimator
  over a \code{OHLC} time series of prices or returns, etc.  By default the
  estimates are trade volume weighted.
}
\examples{
# create time index of one minute intervals over several days
in_dex <- seq(from=as.POSIXct("2015-02-09 09:30:00"),
              to=as.POSIXct("2015-02-28 16:00:00"), by="1 min")
# create xts of random prices
x_ts <- xts(exp(cumsum(0.001*rnorm(length(in_dex)))), order.by=in_dex)
# add trade Volume data
x_ts <- merge(x_ts,
          volume=sample(x=10*(2:18),
            size=length(in_dex), replace=TRUE))
# aggregate to hours OHLC data
oh_lc <- to.period(x=x_ts, period="hours")
# calculate time series of rolling variance and skew estimates
vari_ance <- roll_agg(ohlc=oh_lc)
sk_ew <- roll_agg(ohlc=oh_lc, esti_mator="skew_ohlc")
sk_ew <- sk_ew/(vari_ance)^(1.5)
sk_ew[1, ] <- 0
sk_ew <- na.locf(sk_ew)
}

