% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_count}
\alias{roll_count}
\title{Count the number of consecutive \code{TRUE} elements in a Boolean vector,
and reset the count to zero after every \code{FALSE} element.}
\usage{
roll_count(tseries)
}
\arguments{
\item{\code{tseries}}{A \emph{Boolean vector} of data.}
}
\value{
An \emph{integer vector} of the same length as the argument
  \code{tseries}.
}
\description{
Count the number of consecutive \code{TRUE} elements in a Boolean vector,
and reset the count to zero after every \code{FALSE} element.
}
\details{
The function \code{roll_count()} calculates the number of consecutive
  \code{TRUE} elements in a Boolean vector, and it resets the count to zero
  after every \code{FALSE} element.
  
  For example, the Boolean vector {\code{FALSE}, \code{TRUE}, \code{TRUE},
  \code{FALSE}, \code{FALSE}, \code{TRUE}, \code{TRUE}, \code{TRUE},
  \code{TRUE}, \code{TRUE}, \code{FALSE}}, is translated into {\code{0},
  \code{1}, \code{2}, \code{0}, \code{0}, \code{1}, \code{2}, \code{3},
  \code{4}, \code{5}, \code{0}}.
}
\examples{
\dontrun{
# Calculate the number of consecutive TRUE elements
drop(HighFreq::roll_count(c(FALSE, TRUE, TRUE, FALSE, FALSE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE)))
}
}
