% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_kurtosis}
\alias{roll_kurtosis}
\title{Calculate a \emph{matrix} of kurtosis estimates over a rolling look-back
interval attached at the end points of a \emph{time series} or a
\emph{matrix}.}
\usage{
roll_kurtosis(
  timeser,
  startp = 0L,
  endd = 0L,
  step = 1L,
  lookb = 1L,
  stub = 0L,
  method = "moment",
  confl = 0.75
)
}
\arguments{
\item{\code{timeser}}{A \emph{time series} or a \emph{matrix} of data.}

\item{\code{startp}}{An \emph{integer} vector of start points (the default
is \code{startp = 0}).}

\item{\code{endd}}{An \emph{integer} vector of end points (the default is 
\code{endd = 0}).}

\item{\code{step}}{The number of time periods between the end points (the
default is \code{step = 1}).}

\item{\code{lookb}}{The number of end points in the look-back interval
(the default is \code{lookb = 1}).}

\item{\code{stub}}{An \emph{integer} value equal to the first end point for
calculating the end points (the default is \code{stub = 0}).}

\item{\code{method}}{A \emph{character string} specifying the type of the
kurtosis model (the default is \code{method = "moment"} - see Details).}

\item{\code{confl}}{The confidence level for calculating the quantiles of
returns (the default is \code{confl = 0.75}).}
}
\value{
A \emph{matrix} of kurtosis estimates with the same number of
  columns as the input time series \code{timeser}, and the number of rows
  equal to the number of end points.
}
\description{
Calculate a \emph{matrix} of kurtosis estimates over a rolling look-back
interval attached at the end points of a \emph{time series} or a
\emph{matrix}.
}
\details{
The function \code{roll_kurtosis()} calculates a \emph{matrix} of kurtosis
  estimates over rolling look-back intervals attached at the end points of
  the \emph{time series} \code{timeser}.
  
  The function \code{roll_kurtosis()} performs a loop over the end points,
  and at each end point it subsets the time series \code{timeser} over a
  look-back interval equal to \code{lookb} number of end points.
  
  It passes the subset time series to the function \code{calc_kurtosis()},
  which calculates the kurtosis. See the function \code{calc_kurtosis()} for
  a description of the kurtosis methods.
  
  If the arguments \code{endd} and \code{startp} are not given then it
  first calculates a vector of end points separated by \code{step} time
  periods. It calculates the end points along the rows of \code{timeser}
  using the function \code{calc_endpoints()}, with the number of time
  periods between the end points equal to \code{step} time periods.

  For example, the rolling kurtosis at \code{25} day end points, with a
  \code{75} day look-back, can be calculated using the parameters
  \code{step = 25} and \code{lookb = 3}.

  The function \code{roll_kurtosis()} is implemented in \code{RcppArmadillo}
  \code{C++} code, which makes it several times faster than \code{R} code.
}
\examples{
\dontrun{
# Define time series of returns using package rutils
retp <- na.omit(rutils::etfenv$returns$VTI)
# Define end points and start points
endd <- 1 + HighFreq::calc_endpoints(NROW(retp), step=25)
startp <- HighFreq::calc_startpoints(endd, lookb=3)
# Calculate the rolling kurtosis at 25 day end points, with a 75 day look-back
kurtosisv <- HighFreq::roll_kurtosis(retp, step=25, lookb=3)
# Calculate the rolling kurtosis using R code
kurt_r <- sapply(1:NROW(endd), function(it) {
  HighFreq::calc_kurtosis(retp[startp[it]:endd[it], ])
})  # end sapply
# Compare the kurtosis estimates
all.equal(drop(kurtosisv), kurt_r, check.attributes=FALSE)
# Compare the speed of RcppArmadillo with R code
library(microbenchmark)
summary(microbenchmark(
  Rcpp=HighFreq::roll_kurtosis(retp, step=25, lookb=3),
  Rcode=sapply(1:NROW(endd), function(it) {
    HighFreq::calc_kurtosis(retp[startp[it]:endd[it], ])
  }),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}  # end dontrun

}
