% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_portf}
\alias{roll_portf}
\title{Simulate (backtest) a portfolio momentum strategy, using
\code{RcppArmadillo}.}
\usage{
roll_portf(
  retx,
  retp,
  controll,
  startp,
  endd,
  lambdaf = 0,
  coeff = 1,
  bidask = 0
)
}
\arguments{
\item{\code{retp}}{A \emph{time series} or a \emph{matrix} of asset
returns data.}

\item{\code{retx}}{A \emph{time series} or a \emph{matrix} of excess
returns data (the returns in excess of the risk-free rate).}

\item{\code{controll}}{A \emph{list} of portfolio optimization model
parameters (see Details).}

\item{\code{startp}}{An \emph{integer vector} of start points.}

\item{\code{endd}}{An \emph{integer vector} of end points.}

\item{\code{lambdaf}}{A decay factor which multiplies the past portfolio
weights.  (The default is \code{lambdaf = 0} - no memory.)}

\item{\code{coeff}}{A \emph{numeric} multiplier of the weights.  (The
default is \code{1})}

\item{\code{bidask}}{A \emph{numeric} bid-ask spread (the default is
\code{0})}
}
\value{
A column \emph{vector} of strategy PnLs, with the same length as
  the number of rows of \code{retp}.
}
\description{
Simulate (backtest) a portfolio momentum strategy, using
\code{RcppArmadillo}.
}
\details{
The function \code{roll_portf()} performs a backtest simulation of a
  portfolio momentum strategy over a \emph{vector} of the end
  points \code{endd}.
  
  It performs a loop over the end points \code{endd}, and subsets the
  \emph{matrix} of the excess asset returns \code{retx} along its rows,
  between the corresponding \emph{start point} and the \emph{end point}. 
  
  The \emph{start} and \emph{end} points determine the \emph{in-sample}
  calibration interval.  The \emph{lagg} parameter (from the list controll)
  lags the \emph{start} and \emph{end} points.  The default is \code{lagg =
  1}, which means that the calibration interval ends just before the
  \emph{out-of-sample} test interval starts.
  If \code{lagg=2}, then the calibration interval ends one period before the
  test interval starts.  
  If the \emph{end point} periods are weekly, then the calibration interval
  ends one week before the test interval starts, and the most recent week is
  excluded from the calibration interval. This improves the momentum
  strategy performance, because prices from the most recent week tend to
  revert.
  
  The function \code{roll_portf()} passes the subset matrix of excess
  returns into the function \code{calc_weights()}, which calculates the
  optimal portfolio weights at each \emph{end point}.
  It also passes to \code{calc_weights()} the argument \code{controll},
  which is the list of portfolio optimization parameters.
  See the function \code{calc_weights()} for more details.
  The list of portfolio optimization parameters can be created using the
  function \code{param_portf()}.
  
  The function \code{roll_portf()} then recursively averages the weights
  \eqn{w_i} at the \emph{end point = i} with the weights \eqn{w_{i-1}} from
  the previous \emph{end point = (i-1)}, using the decay factor \code{lambdaf
  = \eqn{\lambda}}:
  \deqn{
    w_i = (1 - \lambda) w_i + \lambda w_{i-1}
  }
  The purpose of averaging the weights is to reduce their variance, and
  improve their out-of-sample performance.  It is equivalent to extending
  the portfolio holding period beyond the time interval between neighboring
  \emph{end points}.
  
  The function \code{roll_portf()} then calculates the out-of-sample strategy
  returns by multiplying the average weights times the future asset returns.
  
  The function \code{roll_portf()} multiplies the out-of-sample strategy
  returns by the coefficient \code{coeff} (with default equal to \code{1}),
  which allows simulating either a trending strategy (if \code{coeff = 1}),
  or a reverting strategy (if \code{coeff = -1}).
  
  The function \code{roll_portf()} calculates the transaction costs by
  multiplying the bid-ask spread \code{bidask} times the absolute
  difference between the current weights minus the weights from the previous
  period. Then it subtracts the transaction costs from the out-of-sample
  strategy PnLs.
  
  The function \code{roll_portf()} returns a \emph{time series} (column
  \emph{vector}) of strategy PnLs, of the same length as the number of
  rows of \code{retp}.
}
\examples{
\dontrun{
# Calculate the ETF daily excess returns
retp <- na.omit(rutils::etfenv$returns[, 1:16])
# riskf is the daily risk-free rate
riskf <- 0.03/260
retx <- retp - riskf
# Define monthly end points without initial warmup period
endd <- rutils::calc_endpoints(retp, interval="months")
endd <- endd[endd > 0]
nrows <- NROW(endd)
# Define 12-month look-back interval and start points over sliding window
lookb <- 12
startp <- c(rep_len(1, lookb-1), endd[1:(nrows-lookb+1)])
# Define return shrinkage and dimension reduction
alphac <- 0.5
dimax <- 3
# Create a list of portfolio optimization parameters
controll <- HighFreq::param_portf(method="maxsharpe", dimax=dimax, alphac=alphac, scalew="sumsq")
# Simulate a monthly portfolio momentum strategy
pnls <- HighFreq::roll_portf(retx, retp, controll=controll, startp=(startp-1), endd=(endd-1))
pnls <- xts::xts(pnls, index(retp))
colnames(pnls) <- "strategy"
# Plot dygraph of strategy
dygraphs::dygraph(cumsum(pnls), 
  main="Cumulative Returns of Max Sharpe Portfolio Strategy")
}  # end dontrun

}
