% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_scale}
\alias{roll_scale}
\title{Perform a rolling scaling (standardization) of the columns of a
\emph{matrix} of data using \emph{RcppArmadillo}.}
\usage{
roll_scale(mat_rix, look_back, use_median = FALSE)
}
\arguments{
\item{mat_rix}{A numeric \emph{matrix} of data.}

\item{look_back}{The length of the look-back interval, equal to the number 
of rows of data used in the scaling.}

\item{use_median}{A \emph{Boolean} argument: if \code{TRUE} then the 
centrality (central tendency) is calculated as the \emph{median} and the 
dispersion is calculated as the \emph{median absolute deviation}
(\emph{MAD}).
If \code{use_median} is \code{FALSE} then the centrality is calculated as 
the \emph{mean} and the dispersion is calculated as the \emph{standard
deviation}. (The default is \code{FALSE})}
}
\value{
A numeric \emph{matrix} with the same dimensions as the input
  argument \code{mat_rix}.
}
\description{
Perform a rolling scaling (standardization) of the columns of a
\emph{matrix} of data using \emph{RcppArmadillo}.
}
\details{
The function \code{roll_scale()} performs a rolling scaling
  (standardization) of the columns of the \code{mat_rix} argument using
  \emph{RcppArmadillo}.
  If the argument \code{use_median} is \code{FALSE} (the default), then it
  performs a similar calculation as the function \code{roll::roll_scale()}.
  If the argument \code{use_median} is \code{TRUE}, then it calculates the
  centrality as the \emph{median} and the dispersion as the \emph{median
  absolute deviation} (\emph{MAD}).
}
\examples{
\dontrun{
mat_rix <- matrix(rnorm(20000), nc=2)
look_back <- 11
rolled_scaled <- roll::roll_scale(data=mat_rix, width=look_back, min_obs=1)
rolled_scaled2 <- roll_scale(mat_rix=mat_rix, look_back=look_back, use_median=FALSE)
all.equal(rolled_scaled[-1, ], rolled_scaled2[-1, ])
}
}
