% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{roll_sharpe}
\alias{roll_sharpe}
\title{Calculate the rolling Sharpe ratio over a rolling lookback window for an
\code{OHLC} time series.}
\usage{
roll_sharpe(oh_lc, win_dow = 11)
}
\arguments{
\item{oh_lc}{an \code{OHLC} time series of prices in \code{xts} format.}

\item{win_dow}{the size of the lookback window, equal to the number of bars
of data used for aggregating the \code{OHLC} prices.}
}
\value{
An \code{xts} time series with a single column and the same number of
  rows as the argument \code{oh_lc}.
}
\description{
Calculate the rolling Sharpe ratio over a rolling lookback window for an
\code{OHLC} time series.
}
\details{
The function \code{roll_sharpe()} calculates the rolling Sharpe
  ratio as the ratio of absolute returns over the lookback window (not
  percentage returns), divided by the average volatility of returns.
}
\examples{
# calculate rolling Sharpe ratio over SPY
sharpe_rolling <- roll_sharpe(oh_lc=SPY, win_dow=10)
}

