% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_var}
\alias{roll_var}
\title{Calculate a \emph{matrix} of variance estimates over a rolling look-back
interval attached at the end points of a \emph{time series} or a
\emph{matrix}.}
\usage{
roll_var(t_series, ste_p = 1L, look_back = 1L)
}
\arguments{
\item{\code{t_series}}{A \emph{time series} or a \emph{matrix}.}

\item{\code{ste_p}}{The number of time periods between the end points.}

\item{\code{look_back}}{The number of end points in the look-back interval.}
}
\value{
A \emph{matrix} with the same number of columns as the input time
  series \code{t_series}, and the number of rows equal to the number of end
  points.
}
\description{
Calculate a \emph{matrix} of variance estimates over a rolling look-back
interval attached at the end points of a \emph{time series} or a
\emph{matrix}.
}
\details{
The function \code{roll_var()} calculates a \emph{matrix} of
  variance estimates over rolling look-back intervals attached at the end
  points of the \emph{time series} \code{t_series}.
  
  The end points are calculated along the rows of \code{t_series} using the
  function \code{calc_endpoints()}, with the number of time periods between
  the end points equal to \code{ste_p}.
  
  At each end point, the variance is calculated over a look-back interval
  equal to \code{look_back} number of end points.
  In the initial warmup period, the variance is calculated over an expanding
  look-back interval.
  
  For example, the rolling variance at \code{25} day end points, with a
  \code{75} day look-back, can be calculated using the parameters
  \code{ste_p = 25} and \code{look_back = 3}.

  The function \code{roll_var()} with the parameter \code{ste_p = 1}
  performs the same calculation as the function \code{roll_var()} from
  package
  \href{https://cran.r-project.org/web/packages/RcppRoll/index.html}{RcppRoll},
  but it's several times faster because it uses \code{RcppArmadillo}
  \code{C++} code.

  The function \code{roll_var()} is implemented in \code{RcppArmadillo}
  \code{C++} code, so it's many times faster than the equivalent \code{R}
  code.
}
\examples{
\dontrun{
# Define time series of returns using package rutils
re_turns <- na.omit(rutils::etf_env$re_turns$VTI)
# Calculate the rolling variance at 25 day end points, with a 75 day look-back
vari_ance <- HighFreq::roll_var(re_turns, ste_p=25, look_back=3)
# Compare the variance estimates over 11-period lookback intervals
all.equal(HighFreq::roll_var(re_turns, look_back=11)[-(1:10), ], 
  drop(RcppRoll::roll_var(re_turns, n=11)), check.attributes=FALSE)
# Compare the speed of RcppArmadillo with RcppRoll
library(microbenchmark)
summary(microbenchmark(
  RcppArmadillo=HighFreq::roll_var(re_turns, look_back=11),
  RcppRoll=RcppRoll::roll_var(re_turns, n=11),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}
}
