% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_var_vec}
\alias{roll_var_vec}
\title{Calculate a \emph{vector} of variance estimates over a rolling look-back
interval for a \emph{vector} or a single-column \emph{time series}, using
\code{RcppArmadillo}.}
\usage{
roll_var_vec(vec_tor, look_back = 11L)
}
\arguments{
\item{\code{vec_tor}}{A \emph{vector} or a single-column \emph{time series}.}

\item{\code{look_back}}{The length of the look-back interval, equal to the
number of \emph{vector} elements used for calculating a single variance
estimate.}
}
\value{
A single-column \emph{vector} with the same number of elements as
  the input argument \code{vec_tor}.
}
\description{
Calculate a \emph{vector} of variance estimates over a rolling look-back
interval for a \emph{vector} or a single-column \emph{time series}, using
\code{RcppArmadillo}.
}
\details{
The function \code{roll_var_vec()} calculates a \code{vec_tor} of
  variance estimates over a rolling look-back interval for a \emph{vector}
  or a single-column \emph{time series}, using \code{RcppArmadillo}.

  The function \code{roll_var_vec()} performs the same calculation as the
  function \code{roll_var()} from package
  \href{https://cran.r-project.org/web/packages/RcppRoll/index.html}{RcppRoll},
  but it's several times faster because it uses \code{RcppArmadillo}.
}
\examples{
\dontrun{
# Create a vector of random returns
re_turns <- rnorm(1e6)
# Compare the variance estimates over 11-period lookback intervals
all.equal(drop(HighFreq::roll_var_vec(re_turns, look_back=11))[-(1:10)], 
  RcppRoll::roll_var(re_turns, n=11))
# Compare the speed of RcppArmadillo with RcppRoll
library(microbenchmark)
summary(microbenchmark(
  RcppArmadillo=HighFreq::roll_var_vec(re_turns, look_back=11),
  RcppRoll=RcppRoll::roll_var(re_turns, n=11),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}
}
