% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{roll_vwap}
\alias{roll_vwap}
\title{Calculate the volume-weighted average price of an \code{OHLC} time series
over a rolling window (lookback period).}
\usage{
roll_vwap(oh_lc, x_ts = Cl(oh_lc), win_dow)
}
\arguments{
\item{oh_lc}{an \code{OHLC} time series of prices in \code{xts} format.}

\item{x_ts}{single-column \code{xts} time series.}

\item{win_dow}{the size of the lookback window, equal to the number of bars
of data used for calculating the average price.}
}
\value{
An \code{xts} time series with a single column and the same number of
  rows as the argument \code{oh_lc}.
}
\description{
Performs the same operation as function \code{VWAP()} from package
\href{https://cran.r-project.org/web/packages/TTR/index.html}{VWAP},
but using vectorized functions, so it's a little faster.
}
\details{
Calculates the volume-weighted average closing price, defined as the
  sum of the prices multiplied by trading volumes in the lookback window,
  divided by the sum of trading volumes in the window.
  If the argument \code{x_ts} is passed in explicitly, then its
  volume-weighted average value over time is calculated.
}
\examples{
# calculate the volume-weighted average closing price (VWAP)
roll_vwap(oh_lc=get("SPY"), win_dow=11)
# calculate the volume-weighted average returns
roll_vwap(oh_lc=get("SPY"), x_ts=re_turns, win_dow=11)
}

