% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{run_moment_ohlc}
\alias{run_moment_ohlc}
\title{Calculate the running estimates over time of the moment of a \code{OHLC} time
series.}
\usage{
run_moment_ohlc(ohlc, mom_fun = "vol_ohlc", calc = "rogers.satchell",
  n = 20, N = 260, vo_lu = TRUE, ...)
}
\arguments{
\item{ohlc}{\code{OHLC} time series of prices.}

\item{mom_fun}{\code{character} string representing function for estimating
the moment.}

\item{n}{\code{numeric} number of periods for averaging of estimates.}

\item{N}{\code{numeric} number of periods in a year (to annualize the estimates).}

\item{vo_lu}{\code{logical} should estimate be weighted by trade volume.}

\item{calc_method}{\code{character} string representing method for estimating
the moment.}
}
\value{
\code{numeric} time series of estimates of the moment.
}
\description{
Calculate the running estimates over time of the moment of a \code{OHLC} time
series.
}
\details{
Calculates a time series of running estimates of the moment of a
  \code{OHLC} time series of prices.  By default the estimates are trade
  volume weighted.
}
\examples{
# create time index of one minute intervals over several days
in_dex <- seq(from=as.POSIXct("2015-02-09 09:30:00"),
              to=as.POSIXct("2015-02-28 16:00:00"), by="1 min")
# create xts of random prices
x_ts <- xts(exp(cumsum(0.001*rnorm(length(in_dex)))), order.by=in_dex)
# add trade Volume data
x_ts <- merge(x_ts,
          volume=sample(x=10*(2:18),
            size=length(in_dex), replace=TRUE))
# aggregate to hours OHLC data
oh_lc <- to.period(x=x_ts, period="hours")
# calculate time series of running volatility and skew estimates
vol_at <- run_moment_ohlc(ohlc=oh_lc)
sk_ew <- run_moment_ohlc(ohlc=oh_lc, mom_fun="skew_ohlc")
sk_ew <- sk_ew/(vol_at)^(1.5)
sk_ew[1, ] <- 0
sk_ew <- na.locf(sk_ew)
}

