% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{run_returns}
\alias{run_returns}
\title{Calculate single period returns from either \code{TAQ} or \code{OHLC} prices.}
\usage{
run_returns(x_ts, col_umn = 4)
}
\arguments{
\item{x_ts}{\code{xts} time series of either \code{TAQ} or \code{OHLC} data.}

\item{col_umn}{the column number to extract from the \code{OHLC} data.
(default is \code{4}, or the \code{close} prices column)}
}
\value{
A single-column \code{xts} time series of returns.
}
\description{
Calculate single period returns from either \code{TAQ} or \code{OHLC} prices.
}
\details{
Calculates single period returns for either \code{TAQ} or 
  \code{OHLC} data, as the ratio of the differenced prices divided by the 
  time index differences. 
  Identifies the \code{x_ts} time series as \code{TAQ} data when it has six
  columns, otherwise assumes it's \code{OHLC} data.
  By default, for \code{OHLC} data, it differences the \code{close} prices, 
  but can also difference other prices depending on the value of
  \code{col_umn}.
}
\examples{
# calculate close to close returns
re_turns <- HighFreq::run_returns(x_ts=SPY)
# calculate open to open returns
re_turns <- HighFreq::run_returns(x_ts=SPY, col_umn=1)
}

