% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{run_skew}
\alias{run_skew}
\title{Calculate time series of skew estimates from a \code{OHLC} time series.}
\usage{
run_skew(oh_lc, calc_method = "rogers.satchell")
}
\arguments{
\item{oh_lc}{an \code{OHLC} time series of prices in \code{xts} format.}

\item{calc_method}{\code{character} string representing method for estimating
skew.}
}
\value{
Time series of skew estimates.
}
\description{
Calculate time series of skew estimates from a \code{OHLC} time series.
}
\details{
The function \code{run_skew()} calculates skew estimates from
  \code{OHLC} prices at each point in time (row).  The methods include
  Garman-Klass and Rogers-Satchell.
}
\examples{
# calculate time series of skew estimates for SPY
sk_ew <- HighFreq::run_skew(SPY)
}

