% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{save_rets}
\alias{save_rets}
\title{Load and scrub multiple days of \code{TAQ} data for a single symbol,
calculate returns, and save the time series to multiple \code{*.RData} files.}
\usage{
save_rets(sym_bol, data_dir = "E:/mktdata/sec/",
  output_dir = "E:/output/data/", period = "minutes")
}
\arguments{
\item{sym_bol}{\code{character} string representing symbol or ticker.}

\item{data_dir}{\code{character} string representing directory containing input \code{*.RData} files.}

\item{output_dir}{\code{character} string representing directory containing output \code{*.RData} files.}
}
\value{
\code{OHLC} time series in \code{xts} format.
}
\description{
Load and scrub multiple days of \code{TAQ} data for a single symbol,
calculate returns, and save the time series to multiple \code{*.RData} files.
}
\details{
The function \code{save_rets} loads multiple days of \code{TAQ}
  data, then scrubs, aggregates, and rbinds it into a \code{OHLC} time
  series, and finally saves it in a single \code{*.RData} file. The
  \code{OHLC} time series is stored in a variable named
  \sQuote{\code{symbol}}, and then it's saved to a file named
  \sQuote{\code{symbol.RData}} in the \sQuote{\code{output_dir}} directory.
  The \code{TAQ} data files are assumed to be stored in separate directories
  for each \sQuote{\code{symbol}}. Each \sQuote{\code{symbol}} has its own
  directory (named \sQuote{\code{symbol}}) in the \sQuote{\code{data_dir}}
  directory. Each \sQuote{\code{symbol}} directory contains multiple daily
  \code{*.RData} files, each file containing one day of \code{TAQ} data.
}
\examples{
\dontrun{
save_rets("SPY")
}
}

