% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{save_rets_OHLC}
\alias{save_rets_OHLC}
\title{Load \code{OHLC} time series data for a single symbol, calculate its returns,
and save them to a single \sQuote{\code{*.RData}} file.}
\usage{
save_rets_OHLC(sym_bol, data_dir = "E:/output/data/",
  output_dir = "E:/output/data/")
}
\arguments{
\item{sym_bol}{\code{character} string representing symbol or ticker.}

\item{data_dir}{\code{character} string representing directory containing input \sQuote{\code{*.RData}} files.}

\item{output_dir}{\code{character} string representing directory containing output \sQuote{\code{*.RData}} files.}
}
\value{
time series of returns and volume in \code{xts} format.
}
\description{
Load \code{OHLC} time series data for a single symbol, calculate its returns,
and save them to a single \sQuote{\code{*.RData}} file.
}
\details{
The function \code{save_rets_OHLC} loads \code{OHLC} time series
  data.  It then calculates returns using function \code{calc_rets}, and
  stores them in a variable named \sQuote{\code{symbol.rets}}, and saves them
  to a file called \sQuote{\code{symbol.rets.RData}}.
  The \code{TAQ} data files are assumed to be stored in separate directories
  for each \sQuote{\code{symbol}}. Each \sQuote{\code{symbol}} has its own
  directory (named \sQuote{\code{symbol}}) in the \sQuote{\code{data_dir}}
  directory. Each \sQuote{\code{symbol}} directory contains multiple daily
  \sQuote{\code{*.RData}} files, each file containing one day of \code{TAQ} data.
}
\examples{
\dontrun{
save_rets_OHLC("SPY")
}
}

