% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{save_scrub_agg}
\alias{save_scrub_agg}
\title{Load, scrub, aggregate, and rbind multiple days of \code{TAQ} data for a
single symbol, and save the \code{OHLC} time series to a single
\sQuote{\code{*.RData}} file.}
\usage{
save_scrub_agg(sym_bol, data_dir = "E:/mktdata/sec/",
  output_dir = "E:/output/data/", win_dow = 51, vol_mult = 2,
  period = "minutes", tzone = "America/New_York")
}
\arguments{
\item{sym_bol}{\code{character} string representing symbol or ticker.}

\item{data_dir}{\code{character} string representing directory containing
input \sQuote{\code{*.RData}} files.}

\item{output_dir}{\code{character} string representing directory containing
output \sQuote{\code{*.RData}} files.}

\item{win_dow}{number of data points for estimating rolling volatility.}

\item{vol_mult}{volatility multiplier.}

\item{period}{aggregation period.}

\item{tzone}{timezone to convert.}
}
\value{
\code{OHLC} time series in \code{xts} format.
}
\description{
Load, scrub, aggregate, and rbind multiple days of \code{TAQ} data for a
single symbol, and save the \code{OHLC} time series to a single
\sQuote{\code{*.RData}} file.
}
\details{
The function \code{save_scrub_agg} loads multiple days of \code{TAQ}
  data, then scrubs, aggregates, and rbinds them into a \code{OHLC} time
  series, and finally saves it to a single \sQuote{\code{*.RData}} file. The
  \code{OHLC} time series is stored in a variable named
  \sQuote{\code{symbol}}, and then it's saved to a file named
  \sQuote{\code{symbol.RData}} in the \sQuote{\code{output_dir}} directory.
  The \code{TAQ} data files are assumed to be stored in separate directories
  for each \sQuote{\code{symbol}}. Each \sQuote{\code{symbol}} has its own
  directory (named \sQuote{\code{symbol}}) in the \sQuote{\code{data_dir}}
  directory. Each \sQuote{\code{symbol}} directory contains multiple daily
  \sQuote{\code{*.RData}} files, each file containing one day of \code{TAQ}
  data.
}
\examples{
\dontrun{
save_scrub_agg("SPY")
}
}

