% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{save_taq}
\alias{save_taq}
\title{Load and scrub multiple days of \emph{TAQ} data for a single symbol, and save
it to multiple \sQuote{\code{*.RData}} files.}
\usage{
save_taq(
  symbol,
  dirin = "E:/mktdata/sec/",
  dirout = "E:/output/data/",
  lookb = 51,
  volm = 2,
  tzone = "America/New_York"
)
}
\value{
a \emph{TAQ} time series in \emph{xts} format.
}
\description{
Load and scrub multiple days of \emph{TAQ} data for a single symbol, and save
it to multiple \sQuote{\code{*.RData}} files.
}
\details{
The function \code{save_taq()} loads multiple days of \emph{TAQ}
  data, scrubs it, and saves the scrubbed TAQ data to individual
  \sQuote{\code{*.RData}} files. It uses the same file names for output as
  the input file names. The \emph{TAQ} data files are assumed to be stored in
  separate directories for each \sQuote{\code{symbol}}. Each
  \sQuote{\code{symbol}} has its own directory (named \sQuote{\code{symbol}})
  in the \sQuote{\code{dirin}} directory.
  Each \sQuote{\code{symbol}} directory contains multiple daily
  \sQuote{\code{*.RData}} files, each file containing one day of \emph{TAQ}
  data.
}
\examples{
\dontrun{
save_taq("SPY")
}

}
