% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{scrub_agg}
\alias{scrub_agg}
\title{Scrub a single day of \code{TAQ} data, aggregate it, and convert to \code{OHLC} format.}
\usage{
scrub_agg(taq_data, vol_window = 51, vol_mult = 2, period = "minutes")
}
\arguments{
\item{taq_data}{\code{TAQ xts} time series.}

\item{vol_window}{number of data points for estimating running volatility.}

\item{vol_mult}{volatility multiplier.}

\item{period}{aggregation period.}
}
\value{
\code{OHLC xts} time series.
}
\description{
Scrub a single day of \code{TAQ} data, aggregate it, and convert to \code{OHLC} format.
}
\details{
The function \code{scrub_agg} performs:
\itemize{
  \item index timezone conversion,
  \item data subset to trading hours,
  \item removal of duplicate time stamps,
  \item scrubbing of quotes with suspect bid-offer spreads,
  \item scrubbing of quotes with suspect price jumps,
  \item cbinding of mid prices with volume data,
  \item aggregation to OHLC using \code{to.period} from \code{xts},
}
Valid 'period' character strings include: "minutes", "3 min", "5 min", "10
min", "15 min", "30 min", and "hours". The time index of the output time
series is rounded up to the next integer multiple of 'period'.
}
\examples{
# create time index of one second intervals for a single day
in_dex <- seq(from=as.POSIXct("2015-02-09 09:30:00"), to=as.POSIXct("2015-02-09 16:00:00"), by="1 sec")
# create xts of random TAQ prices
x_ts <- xts(cumsum(rnorm(length(in_dex))), order.by=in_dex)
# create vector of random bid-offer prices
bid_offer <- abs(rnorm(length(in_dex)))/10
# create TAQ data using cbind
taq_data <- cbind(x_ts-bid_offer, x_ts+bid_offer)
# add Trade.Price
taq_data <- cbind(taq_data, x_ts+rnorm(length(in_dex))/10)
# add Volume
taq_data <- cbind(taq_data, sample(x=10*(2:18), size=length(in_dex), replace=TRUE))
colnames(taq_data) <- c("Bid.Price", "Ask.Price", "Trade.Price", "Volume")
# aggregate to ten minutes OHLC data
ohlc_data <- scrub_agg(taq_data, period="10 min")
chartSeries(ohlc_data, name=sym_bol, theme=chartTheme("white"))
}

