% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_df}
\alias{sim_df}
\title{Simulate a \emph{Dickey-Fuller} process using \emph{Rcpp}.}
\usage{
sim_df(prici, priceq, theta, coeff, innov)
}
\arguments{
\item{\code{prici}}{The initial price.}

\item{\code{priceq}}{The equilibrium price.}

\item{\code{theta}}{The strength of mean reversion.}

\item{\code{coeff}}{A single-column \emph{matrix} of \emph{autoregressive}
coefficients.}

\item{\code{innov}}{A single-column \emph{matrix} of innovations (random
numbers).}
}
\value{
A single-column \emph{matrix} of simulated prices, with the same
  number of rows as the argument \code{innov}.
}
\description{
Simulate a \emph{Dickey-Fuller} process using \emph{Rcpp}.
}
\details{
The function \code{sim_df()} simulates the following \emph{Dickey-Fuller}
  process:
  \deqn{
    r_i = \theta \, (\mu - p_{i-1}) + \varphi_1 r_{i-1} + \ldots + \varphi_n r_{i-n} + \xi_i
  }
  \deqn{
    p_i = p_{i-1} + r_i
  }
  Where \eqn{r_i} and \eqn{p_i} are the simulated returns and prices,
  \eqn{\theta} and \eqn{\mu} are the \emph{Ornstein-Uhlenbeck} parameters,
  \eqn{\varphi_i} are the \emph{autoregressive} coefficients, and
  \eqn{\xi_i} are the normal \emph{innovations}.
  The recursion starts with: \eqn{r_1 = \xi_1} and \eqn{p_1 = init\_price}.

  The \emph{Dickey-Fuller} process is a combination of an
  \emph{Ornstein-Uhlenbeck} process and an \emph{autoregressive} process.
  The order \eqn{n} of the \emph{autoregressive} process \eqn{AR(n)}, is
  equal to the number of rows of the \emph{autoregressive} coefficients
  \code{coeff}.

  The function \code{sim_df()} simulates the \emph{Dickey-Fuller}
  process using fast \emph{Rcpp} \code{C++} code.

  The function \code{sim_df()} returns a single-column \emph{matrix}
  representing the \emph{time series} of prices.
}
\examples{
\dontrun{
# Define the Ornstein-Uhlenbeck model parameters
prici <- 1.0
priceq <- 2.0
thetav <- 0.01
# Define AR coefficients
coeff <- matrix(c(0.1, 0.3, 0.5))
# Calculate matrix of standard normal innovations
innov <- matrix(rnorm(1e3, sd=0.01))
# Simulate Dickey-Fuller process using Rcpp
prices <- HighFreq::sim_df(prici=prici, priceq=priceq, theta=thetav, coeff=coeff, innov=innov)
plot(prices, t="l", main="Simulated Dickey-Fuller Prices")
}

}
