% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_garch}
\alias{sim_garch}
\title{Simulate a \emph{GARCH} process using \emph{Rcpp}.}
\usage{
sim_garch(omega, alpha, beta, innov)
}
\arguments{
\item{\code{omega}}{Parameter proportional to the long-term average level
of variance.}

\item{\code{alpha}}{The weight associated with recent realized variance
updates.}

\item{\code{beta}}{The weight associated with the past variance estimates.}

\item{\code{innov}}{A \emph{vector} of innovations (random numbers).}
}
\value{
A \emph{matrix} with two columns: the simulated returns and
  variance, and with the same number of rows as the length of the argument 
  \code{innov}.
}
\description{
Simulate a \emph{GARCH} process using \emph{Rcpp}.
}
\details{
The function \code{sim_garch()} simulates a \emph{GARCH} process using
  fast \emph{Rcpp} \code{C++} code.
}
\examples{
\dontrun{
# Define the GARCH model parameters
ome_ga <- 0.01
al_pha <- 0.5
be_ta <- 0.2
# Simulate the GARCH process using Rcpp
garch_rcpp <- sim_garch(omega=ome_ga, alpha=al_pha, beta=be_ta, innov=rnorm(10000))
}

}
