% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{v_wap}
\alias{v_wap}
\title{Calculate the volume-weighted average price of an \code{OHLC} time series
over a sliding window (lookback period).}
\usage{
v_wap(x_ts, win_dow)
}
\arguments{
\item{x_ts}{an \code{xts} time series containing one or more columns of data.}

\item{win_dow}{an integer specifying the number of lookback periods.}
}
\value{
\code{xts} time series with a single column containing the VWAP of
  the close prices, with the same number of rows as the input \code{xts} series.
}
\description{
Performs the same operation as function \code{VWAP()} from package
\href{https://cran.r-project.org/web/packages/TTR/index.html}{VWAP},
but using vectorized functions, so it's a little faster.
}
\details{
The volume-weighted average price (VWAP) over a period is defined as
  the sum of the prices multiplied by trading volumes, divided by the total
  trading volume in that period.
}
\examples{
# create xts time series
x_ts <- xts(x=rnorm(1000), order.by=(Sys.time()-3600*(1:1000)))
v_wap(x_ts=get("SPY"), win_dow=11)
}

