% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{vari_ance}
\alias{vari_ance}
\title{Calculate a time series of variance estimates for an \code{OHLC} time series.}
\usage{
vari_ance(ohlc, calc_method = "garman.klass_yz")
}
\arguments{
\item{ohlc}{\code{OHLC} time series of prices.}

\item{calc_method}{\code{character} string representing method for estimating
  variance.  The methods include: 
\itemize{
  \item "close" close to close, 
  \item "garman.klass" Garman-Klass, 
  \item "garman.klass_yz" Garman-Klass with account for close-to-open jumps, 
  \item "rogers.satchell" Rogers-Satchell, 
  \item "yang.zhang" Yang-Zhang, 
}}
}
\value{
A single-column \code{xts} time series of variance estimates, with
  the same number of rows as the input time series.
}
\description{
Calculates variance estimates for each bar of \code{OHLC} prices at each
point in time (row), using the squared differences of \code{OHLC} prices at
each point in time.
}
\details{
Performs a similar operation as function \code{volatility()} from
  package \href{https://cran.r-project.org/web/packages/TTR/index.html}{TTR},
  but without calculating a running sum using \code{runSum()}.  It's also a 
  little faster because it performs less data validation.
}
\examples{
# create time index of one second intervals for a single day
in_dex <- seq(from=as.POSIXct("2015-02-09 09:30:00"),
              to=as.POSIXct("2015-02-09 16:00:00"), by="1 sec")
# create xts of random prices
x_ts <- xts(cumsum(rnorm(length(in_dex))), order.by=in_dex)
# aggregate to minutes OHLC data
oh_lc <- to.period(x=x_ts, period="minutes")
# calculate variance estimates
vari_ance <- vari_ance(oh_lc)
# calculate variance estimates for SPY
vari_ance <- vari_ance(SPY, calc_method="yang.zhang")
}

