% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{agg_ohlc}
\alias{agg_ohlc}
\title{Aggregate a time series of data into a single bar of \emph{OHLC} data.}
\usage{
agg_ohlc(timeser)
}
\arguments{
\item{\code{timeser}}{A \emph{time series} or a \emph{matrix} with multiple
columns of data.}
}
\value{
A \emph{matrix} containing a single row, with the \emph{open},
  \emph{high}, \emph{low}, and \emph{close} values, and also the total
  \emph{volume} (if provided as either the second or fifth column of
  \code{timeser}).
}
\description{
Aggregate a time series of data into a single bar of \emph{OHLC} data.
}
\details{
The function \code{agg_ohlc()} aggregates a time series of data into a
  single bar of \emph{OHLC} data. It can accept either a single column of
  data or four columns of \emph{OHLC} data.
  It can also accept an additional column containing the trading volume.
  
The function \code{agg_ohlc()} calculates the \emph{open} value as equal to
the \emph{open} value of the first row of \code{timeser}.
  The \emph{high} value as the maximum of the \emph{high} column of
  \code{timeser}.
  The \emph{low} value as the minimum of the \emph{low} column of
  \code{timeser}.
  The \emph{close} value as the \emph{close} of the last row of
  \code{timeser}.
  The \emph{volume} value as the sum of the \emph{volume} column of
  \code{timeser}.

  For a single column of data, the \emph{open}, \emph{high}, \emph{low}, and
  \emph{close} values are all the same.
}
\examples{
\dontrun{
# Define matrix of OHLC data
ohlc <- coredata(rutils::etfenv$VTI[, 1:5])
# Aggregate to single row matrix
ohlcagg <- HighFreq::agg_ohlc(ohlc)
# Compare with calculation in R
all.equal(drop(ohlcagg),
  c(ohlc[1, 1], max(ohlc[, 2]), min(ohlc[, 3]), ohlc[NROW(ohlc), 4], sum(ohlc[, 5])), 
  check.attributes=FALSE)
}

}
