% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{agg_ohlc}
\alias{agg_ohlc}
\title{Aggregate an \emph{OHLC} time series to a single \emph{OHLC} bar.}
\usage{
agg_ohlc(oh_lc)
}
\arguments{
\item{\code{oh_lc}}{A \emph{matrix} or \emph{time series} with four or
five columns of \emph{OHLC} data.}
}
\value{
A \emph{single row matrix} with the \emph{open}, \emph{high},
  \emph{low}, and \emph{close} values, and also the total \emph{volume} (if
  provided in the fifth column of \code{oh_lc}).
}
\description{
Aggregate an \emph{OHLC} time series to a single \emph{OHLC} bar.
}
\details{
The function \code{agg_ohlc()} calculates the \emph{open} value as
  equal to the \emph{open} of the first row of \code{oh_lc}.
  The \emph{high} value as the maximum of the \emph{high} column of
  \code{oh_lc}.
  The \emph{low} value as the minimum of the \emph{low} column of
  \code{oh_lc}.
  The \emph{close} value as the \emph{close} of the last row of
  \code{oh_lc}.
  The \emph{volume} value as the sum of the \emph{volume} column of
  \code{oh_lc}.
}
\examples{
\dontrun{
# Define matrix of OHLC data
oh_lc <- coredata(rutils::etf_env$VTI[, 1:5])
# Aggregate to single row matrix
ohlc_agg <- HighFreq::agg_ohlc(oh_lc)
# Compare with calculation in R
all.equal(drop(ohlc_agg),
  c(oh_lc[1, 1], max(oh_lc[, 2]), min(oh_lc[, 3]), oh_lc[NROW(oh_lc), 4], sum(oh_lc[, 5])), 
  check.attributes=FALSE)
}

}
