% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_eigen}
\alias{calc_eigen}
\title{Calculate the eigen decomposition of the covariance matrix of returns using
\emph{RcppArmadillo}.}
\usage{
calc_eigen(mat_rix)
}
\arguments{
\item{mat_rix}{A numeric \emph{matrix} of returns data.}
}
\value{
A list with two elements: a numeric \emph{vector} of eigenvalues 
  (named "values"), and a numeric \emph{matrix} of eigenvectors (named
  "vectors").
}
\description{
Calculate the eigen decomposition of the covariance matrix of returns using
\emph{RcppArmadillo}.
}
\details{
The function \code{calc_eigen()} first calculates the covariance 
  matrix of the \code{mat_rix}, and then calculates its eigen decomposition.
}
\examples{
\dontrun{
# Create random matrix
mat_rix <- matrix(rnorm(500), nc=5)
# Calculate eigen decomposition
ei_gen <- HighFreq::calc_eigen(scale(mat_rix, scale=FALSE))
# Calculate PCA
pc_a <- prcomp(mat_rix)
# Compare PCA with eigen decomposition
all.equal(pc_a$sdev^2, drop(ei_gen$values))
all.equal(abs(unname(pc_a$rotation)), abs(ei_gen$vectors))
}
}
