% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_eigen}
\alias{calc_eigen}
\title{Calculate the eigen decomposition of the covariance \emph{matrix} of returns
data using \code{RcppArmadillo}.}
\usage{
calc_eigen(tseries)
}
\arguments{
\item{\code{tseries}}{A \emph{time series} or \emph{matrix} of returns
data.}
}
\value{
A list with two elements: a \emph{vector} of eigenvalues 
  (named "values"), and a \emph{matrix} of eigenvectors (named
  "vectors").
}
\description{
Calculate the eigen decomposition of the covariance \emph{matrix} of returns
data using \code{RcppArmadillo}.
}
\details{
The function \code{calc_eigen()} first calculates the covariance
  \emph{matrix} of \code{tseries}, and then calculates the eigen
  decomposition of the covariance \emph{matrix}.
}
\examples{
\dontrun{
# Create matrix of random data
datav <- matrix(rnorm(5e6), nc=5)
# Calculate eigen decomposition
eigend <- HighFreq::calc_eigen(scale(datav, scale=FALSE))
# Calculate PCA
pcad <- prcomp(datav)
# Compare PCA with eigen decomposition
all.equal(pcad$sdev^2, drop(eigend$values))
all.equal(abs(unname(pcad$rotation)), abs(eigend$vectors))
# Compare the speed of Rcpp with R code
summary(microbenchmark(
  Rcpp=HighFreq::calc_eigen(datav),
  Rcode=prcomp(datav),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
