% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_hurst}
\alias{calc_hurst}
\title{Calculate the Hurst exponent from the volatility ratio of aggregated returns.}
\usage{
calc_hurst(tseries, step)
}
\arguments{
\item{\code{tseries}}{A \emph{time series} or a \emph{matrix} of prices.}

\item{\code{step}}{The number of time periods in each interval between
neighboring end points.}
}
\value{
The Hurst exponent calculated from the volatility ratio of
  aggregated returns.
}
\description{
Calculate the Hurst exponent from the volatility ratio of aggregated returns.
}
\details{
The function \code{calc_hurst()} calculates the Hurst exponent from the
  ratios of the volatilities of aggregated returns.

  The aggregated volatility \eqn{\sigma_t} increases with the length of the
  aggregation interval \eqn{\Delta t}.
  The aggregated volatility increases as the length of the aggregation
  interval \eqn{\Delta t} raised to the power of the \emph{Hurst exponent}
  \eqn{H}:
    \deqn{
      \sigma_t = \sigma {\Delta t}^H
    }
  Where \eqn{\sigma} is the daily return volatility.

  The \emph{Hurst exponent} \eqn{H} is equal to the logarithm of the ratio
  of the volatilities divided by the logarithm of the time interval
  \eqn{\Delta t}:
    \deqn{
      H = \frac{\log{\sigma_t} - \log{\sigma}}{\log{\Delta t}}
    }

  The function \code{calc_hurst()} calls the function \code{calc_var_ag()}
  to calculate the aggregated variance \eqn{\sigma^2_t}.
}
\examples{
\dontrun{
# Calculate the log prices
prices <- na.omit(rutils::etfenv$prices[, c("XLP", "VTI")])
prices <- log(prices)
# Calculate the Hurst exponent from 21 day aggregations
calc_hurst(prices, step=21)
}

}
