% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_hurst}
\alias{calc_hurst}
\title{Calculate the Hurst exponent from the volatility ratio of aggregated returns.}
\usage{
calc_hurst(timeser, aggv)
}
\arguments{
\item{\code{timeser}}{A \emph{time series} or a \emph{matrix} of log prices.}

\item{\code{aggv}}{A \emph{vector} of aggregation intervals.}
}
\value{
The Hurst exponent calculated from the volatility ratio of
  aggregated returns.  If \code{timeser} contains multiple columns, then the
  function \code{calc_hurst()} returns a single-row matrix of Hurst
  exponents.
}
\description{
Calculate the Hurst exponent from the volatility ratio of aggregated returns.
}
\details{
The function \code{calc_hurst()} calculates the Hurst exponent from the
  ratios of the volatilities of aggregated returns.
  
  An aggregation interval is equal to the number of time periods between the
  neighboring aggregation end points.

  The aggregated volatility \eqn{\sigma_t} increases with the length of the
  aggregation interval \eqn{\Delta t}.
  The aggregated volatility increases as the length of the aggregation
  interval \eqn{\Delta t} raised to the power of the \emph{Hurst exponent}
  \eqn{H}:
    \deqn{
      \sigma_t = \sigma {\Delta t}^H
    }
  Where \eqn{\sigma} is the daily return volatility.

  For a single aggregation interval \eqn{\Delta t}, the \emph{Hurst
  exponent} \eqn{H} is equal to the logarithm of the ratio of the
  volatilities divided by the logarithm of the aggregation interval
  \eqn{\Delta t}:
    \deqn{
      H = \frac{\log{\sigma_t} - \log{\sigma}}{\log{\Delta t}}
    }

  For a \emph{vector} of aggregation intervals \eqn{\Delta t_i}, the
  \emph{Hurst exponent} \eqn{H} is equal to the regression slope between the
  logarithms of the aggregated volatilities \eqn{\sigma_i} versus the
  logarithms of the aggregation intervals \eqn{\Delta t_i}:
    \deqn{
      H = \frac{\code{cov}(\log{\sigma_i}, \log{\Delta t_i})}{\code{var}(\log{\Delta t_i})}
    }

  The function \code{calc_hurst()} calls the function \code{calc_var_ag()}
  to calculate the variance of aggregated returns \eqn{\sigma^2_t}.
}
\examples{
\dontrun{
# Calculate the log prices
closep <- na.omit(rutils::etfenv$prices[, c("XLP", "VTI")])
closep <- log(closep)
# Calculate the Hurst exponents for a 21 day aggregation interval
HighFreq::calc_hurst(closep, aggv=21)
# Calculate the Hurst exponents for a vector of aggregation intervals
aggv <- seq.int(from=3, to=35, length.out=9)^2
HighFreq::calc_hurst(closep, aggv=aggv)
}  # end dontrun

}
