% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_var_ohlc_ag}
\alias{calc_var_ohlc_ag}
\title{Calculate the variance of aggregated \emph{OHLC} prices using different
price range estimators.}
\usage{
calc_var_ohlc_ag(
  ohlc,
  step = 1L,
  method = "yang_zhang",
  close_lag = 0L,
  scale = TRUE,
  index = 0L
)
}
\arguments{
\item{\code{ohlc}}{A \emph{time series} or a \emph{matrix} of \emph{OHLC}
prices.}

\item{\code{step}}{The number of periods in each interval between
neighboring end points.}

\item{\code{method}}{A \emph{character} string representing the price range
estimator for calculating the variance.  The estimators include:
\itemize{
  \item "close" close-to-close estimator,
  \item "rogers_satchell" Rogers-Satchell estimator,
  \item "garman_klass" Garman-Klass estimator,
  \item "garman_klass_yz" Garman-Klass with account for close-to-open price jumps,
  \item "yang_zhang" Yang-Zhang estimator,
 }
 (The default is the \code{method = "yang_zhang"}.)}

\item{\code{close_lag}}{A \emph{vector} with the lagged \emph{close} prices
of the \emph{OHLC time series}.  This is an optional argument. (The
default is \code{close_lag = 0}).}

\item{\code{scale}}{\emph{Boolean} argument: Should the returns be divided
by the time index, the number of seconds in each period? (The default is
\code{scale = TRUE}).}

\item{\code{index}}{A \emph{vector} with the time index of the \emph{time
series}.  This is an optional argument (the default is \code{index = 0}).}
}
\value{
The variance of aggregated \emph{OHLC} prices.
}
\description{
Calculate the variance of aggregated \emph{OHLC} prices using different
price range estimators.
}
\details{
The function \code{calc_var_ohlc_ag()} calculates the variance of
  \emph{OHLC} prices aggregated over end points.

  It first calculates the end points spaced apart by the number of periods
  equal to the argument \code{step}.  Then it aggregates the \emph{OHLC}
  prices to the end points. Finally it calculates the variance of the
  aggregated \emph{OHLC} prices.

  If there are extra periods that don't fit over the length of \code{ohlc},
  then \code{calc_var_ohlc_ag()} loops over all possible stub intervals,
  it calculates all the corresponding variance values, and it averages
  them.

  For example, if the number of rows of \code{ohlc} is equal to \code{20},
  and \code{step=3} then \code{6} end points fit over the length of
  \code{ohlc}, and there are \code{2} extra periods that must fit into
  stubs, either at the beginning or at the end (or both).

  The aggregated volatility \eqn{\sigma_t} scales (increases) with the
  length of the aggregation interval \eqn{\Delta t} raised to the power of
  the \emph{Hurst exponent} \eqn{H}:
    \deqn{
      \sigma_t = \sigma {\Delta t}^H
    }
  Where \eqn{\sigma} is the daily return volatility.

  The function \code{calc_var_ohlc_ag()} can therefore be used to calculate the
  \emph{Hurst exponent} from the volatility ratio.
}
\examples{
\dontrun{
# Calculate the log ohlc prices
oh_lc <- log(rutils::etf_env$VTI)
# Calculate the daily variance of percentage returns
calc_var_ohlc_ag(oh_lc, step=1)
# Calculate the variance of returns aggregated over 21 days
calc_var_ohlc_ag(oh_lc, step=21)
# The variance over 21 days is approximately 21 times the daily variance
21*calc_var_ohlc_ag(oh_lc, step=1)
}

}
