% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_var_vec}
\alias{calc_var_vec}
\title{Calculate the variance of a \emph{vector} or a single-column \emph{time
series} using \code{RcppArmadillo}.}
\usage{
calc_var_vec(re_turns)
}
\arguments{
\item{\code{re_turns}}{A \emph{vector} or a single-column \emph{time series}.}
}
\value{
A \emph{numeric} value equal to the variance of the \emph{vector}.
}
\description{
Calculate the variance of a \emph{vector} or a single-column \emph{time
series} using \code{RcppArmadillo}.
}
\details{
The function \code{calc_var_vec()} calculates the variance of a
  \emph{vector} using \code{RcppArmadillo}, so it's significantly faster
  than the \code{R} function \code{var()}.
}
\examples{
\dontrun{
# Create a vector of random returns
re_turns <- rnorm(1e6)
# Compare calc_var_vec() with standard var()
all.equal(HighFreq::calc_var_vec(re_turns), 
  var(re_turns))
# Compare the speed of RcppArmadillo with R code
library(microbenchmark)
summary(microbenchmark(
  rcpp=HighFreq::calc_var_vec(re_turns),
  rcode=var(re_turns),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
