% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate the maximum Sharpe ratio portfolio weights for the matrix returns,
using the regularized inverse of the covariance matrix, and
\emph{RcppArmadillo}.}
\usage{
calc_weights(re_turns, alpha_s, alphas_b, max_eigen)
}
\arguments{
\item{re_turns}{A numeric \emph{matrix} of returns data.}

\item{vec_tor}{A numeric \emph{vector} of data.}

\item{look_back}{The length of the look-back interval, equal to the number of rows
of data used for calculating the variance.}
}
\value{
A numeric \emph{vector} of the same length as the argument \code{vec_tor}.
}
\description{
Calculate the maximum Sharpe ratio portfolio weights for the matrix returns,
using the regularized inverse of the covariance matrix, and
\emph{RcppArmadillo}.
}
\details{
The function \code{calc_weights()} calculates a \emph{vector} of rolling 
  variance estimates, from over a \emph{vector} of returns, using \emph{Rcpp}.
}
\examples{
\dontrun{
# create minutely OHLC time series of random prices
vec_tor <- HighFreq::random_ohlc()
# calculate variance estimates for vec_tor over a 21 period interval
var_rolling <- HighFreq::calc_weights(vec_tor, look_back=21)
# calculate variance estimates for SPY
var_rolling <- HighFreq::calc_weights(HighFreq::SPY, calc_method="yang_zhang")
# calculate SPY variance without accounting for overnight jumps
var_rolling <- HighFreq::calc_weights(HighFreq::SPY, calc_method="rogers_satchell")
}
}
