% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diff_it}
\alias{diff_it}
\title{Calculate the row differences of a \emph{matrix} or a \emph{time
series} using \emph{RcppArmadillo}.}
\usage{
diff_it(mat_rix, lagg = 1L, padd = FALSE)
}
\arguments{
\item{\code{mat_rix}}{A \emph{matrix} or \emph{time series}.}

\item{\code{lagg}}{An \emph{integer} equal to the number of rows (time
periods) to lag when calculating the differences (the default is
\code{lagg=1}).}

\item{\code{padd}}{\emph{Boolean} argument: Should the output \emph{matrix}
be padded (extended) with zeros, in order to return a \emph{matrix} with
the same number of rows as the input? (the default is \code{padd=FALSE})}
}
\value{
A \emph{matrix} containing the differences of the input
  \emph{matrix}.
}
\description{
Calculate the row differences of a \emph{matrix} or a \emph{time
series} using \emph{RcppArmadillo}.
}
\details{
The function \code{diff_it()} calculates the differences between
  the rows of the input \emph{matrix} or \emph{time series} and its lagged
  version. The lagged version has its rows shifted down by the number equal
  to \code{lagg} rows.
  
  The argument \code{lagg} specifies the number of lags applied to the rows
  of the lagged version. For example, if \code{lagg=3} then the lagged
  version will have its rows shifted down by \code{3} rows, and the
  differences will be taken between each row minus the row three time
  periods before it (in the past). The default is \code{lagg=1}.

  The argument \code{padd} specifies whether the output \emph{matrix} should
  be padded (extended) with rows of zeros at the beginning, in order to
  return a \emph{matrix} with the same number of rows as the input.  The
  default is \code{padd=FALSE}. The padding operation is time-consuming, so
  that \code{padd=FALSE} can be twice as fast as \code{padd=TRUE}.
  
  The function \code{diff_it()} is implemented in \code{RcppArmadillo}
  code, which makes it slightly faster than \code{R} code.
}
\examples{
\dontrun{
# Create a matrix of random returns
re_turns <- matrix(rnorm(5e6), nc=5)
# Compare diff_it() with rutils::diff_it()
all.equal(HighFreq::diff_it(re_turns, padd=TRUE),
  rutils::diff_it(re_turns))
# Compare the speed of RcppArmadillo with R code
library(microbenchmark)
summary(microbenchmark(
  rcpp=HighFreq::diff_it(re_turns, padd=TRUE),
  rcode=rutils::diff_it(re_turns),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
