% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{do_call_rbind}
\alias{do_call_rbind}
\title{Recursively "rbind" a list of objects.}
\usage{
do_call_rbind(list_var)
}
\arguments{
\item{list_var}{A list of \code{vectors}, \code{matrices}, \code{data
frames}, or \code{time series}.}
}
\value{
a single \code{vector}, \code{matrix}, \code{data frame}, or
  \code{time series}.
}
\description{
Performs the same operation as \code{do.call(rbind, list_var)}, but using
recursion, which is much faster and uses less memory. This is the same
function as '\code{\link[qmao]{do.call.rbind}}' from package 'qmao'.
}
\details{
Performing a \code{do.call(rbind, x_ts)} on a list of 'xts' time
  series is very slow, and often causes an "out of memory" error.
}
\examples{
# create xts time series
x_ts <- xts(x=rnorm(1000), order.by=(Sys.time()-3600*(1:1000)))
# split time series into daily list
list_xts <- split(x_ts, "days")
# rbind the list back into a time series and compare with the original
identical(x_ts, do_call_rbind(list_xts))
}

