% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{extreme_values}
\alias{extreme_values}
\title{Identify extreme values in a univariate xts time series.}
\usage{
extreme_values(time_series, vol_window = 51, thresh_old = 2)
}
\arguments{
\item{time_series}{univariate \code{xts} time series.}

\item{vol_window}{number of data points for estimating running volatility.}

\item{thresh_old}{level of prices above which values are considered extreme.}
}
\value{
\code{logical vector}.
}
\description{
Calculates running volatility defined as a quantile over a sliding window.
Identifies extreme values as those that exceed the scaled running volatility.
}
\details{
Extreme values are those that exceed a fixed multiple of the running
  volatility.
}
\examples{
# create xts time series
x_ts <- xts(x=rnorm(1000), order.by=(Sys.time()-3600*(1:1000)))
# scrub extreme values
x_ts <- x_ts[!extreme_values(x_ts, thresh_old=1)]
}

