% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{extreme_values}
\alias{extreme_values}
\title{Identify extreme values in a univariate \code{xts} time series.}
\usage{
extreme_values(x_ts, win_dow = 51, vol_mult = 2)
}
\arguments{
\item{x_ts}{univariate \code{xts} time series.}

\item{win_dow}{number of data points for estimating rolling volatility.}

\item{vol_mult}{volatility multiplier.}
}
\value{
\code{Boolean} vector with the same number of rows as input time
  series.
}
\description{
Identifies extreme values as those that exceed a multiple of the rolling
volatility.
}
\details{
Calculates rolling volatility as a quantile of values over a sliding
  window. Extreme values are those that exceed the product of the volatility
  multiplier times the rolling volatility. Extreme values are the very tips
  of the tails when the distribution of values becomes very fat-tailed. The
  volatility multiplier \code{vol_mult} controls the threshold at which
  values are identified as extreme. Smaller volatility multiplier values will
  cause more values to be identified as extreme.
}
\examples{
# create xts time series
x_ts <- xts(x=rnorm(1000), order.by=(Sys.time()-3600*(1:1000)))
# scrub extreme values
x_ts <- x_ts[!extreme_values(x_ts, vol_mult=1)]
}

