% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{ohlc_returns}
\alias{ohlc_returns}
\title{Calculate single period percentage returns from either \emph{TAQ} or
\emph{OHLC} prices.}
\usage{
ohlc_returns(xtsv, lagg = 1, colnum = 4, scalit = TRUE)
}
\arguments{
\item{\code{xtsv}}{An \emph{xts} time series of either \emph{TAQ} or \emph{OHLC} data.}

\item{\code{lagg}}{An integer equal to the number of time periods of lag. (default
is 1)}

\item{\code{colnum}}{The column number to extract from the \emph{OHLC} data.
(default is \code{4}, or the \emph{Close} prices column)}

\item{\code{scalit}}{\emph{Boolean} argument: should the returns be divided by the
number of seconds in each period? (default is \code{TRUE})}
}
\value{
A single-column \emph{xts} time series of returns.
}
\description{
Calculate single period percentage returns from either \emph{TAQ} or
\emph{OHLC} prices.
}
\details{
The function \code{ohlc_returns()} calculates the percentage returns
  for either \emph{TAQ} or \emph{OHLC} data, defined as the difference of log
  prices.  Multi-period returns can be calculated by setting the \code{lag}
  parameter to values greater than \code{1} (the default).

  If \code{scalit} is \code{TRUE} (the default), then the returns are divided
  by the differences of the time index (which scales the returns to units of
  returns per second.)

  The time index of the \code{xtsv} time series is assumed to be in
  \emph{POSIXct} format, so that its internal value is equal to the number of
  seconds that have elapsed since the \emph{epoch}.

  If \code{scalit} is \code{TRUE} (the default), then the returns are
  expressed in the scale of the time index of the \code{xtsv} time series.
  For example, if the time index is in seconds, then the returns are given in
  units of returns per second.  If the time index is in days, then the
  returns are equal to the returns per day.

  The function \code{ohlc_returns()} identifies the \code{xtsv} time series as
  \emph{TAQ} data when it has six columns, otherwise assumes it's \emph{OHLC}
  data. By default, for \emph{OHLC} data, it differences the \emph{Close}
  prices, but can also difference other prices depending on the value of
  \code{colnum}.
}
\examples{
# Calculate secondly returns from TAQ data
returns <- HighFreq::ohlc_returns(xtsv=HighFreq::SPY_TAQ)
# Calculate close to close returns
returns <- HighFreq::ohlc_returns(xtsv=HighFreq::SPY)
# Calculate open to open returns
returns <- HighFreq::ohlc_returns(xtsv=HighFreq::SPY, colnum=1)
}
