% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{random_TAQ}
\alias{random_TAQ}
\title{Calculate a random \code{TAQ} time series of prices and trading volumes, in 
\code{xts} format.}
\usage{
random_TAQ(in_dex = seq(from = as.POSIXct(paste(Sys.Date() - 3, "09:30:00")),
  to = as.POSIXct(paste(Sys.Date() - 1, "16:00:00")), by = "1 sec"),
  bid_offer = 0.001, ...)
}
\arguments{
\item{in_dex}{time index for the \code{TAQ} time series.}

\item{bid_offer}{the bid-offer spread expressed as a fraction of the prices. 
The default value is equal to 0.001 (10bps).}
}
\value{
an \code{xts} time series, with time index equal to the input 
  \code{in_dex} time index, and with four columns containing the bid, ask,
  and trade prices, and the trade volume.
}
\description{
Calculate a \code{TAQ} time series of prices and trading volumes, using
random log-normal prices and a time index.
}
\details{
The function \code{random_TAQ()} calculates an \code{xts} time
  series with four columns containing random log-normal prices: the bid, ask,
  and trade prices, and the trade volume.
  If \code{in_dex} isn't supplied as an argument, then by default it's
  equal to the secondly index over the two previous calendar days.
}
\examples{
# create secondly TAQ time series of random prices
ta_q <- HighFreq::random_TAQ()
# create random TAQ time series from SPY index
ta_q <- HighFreq::random_TAQ(in_dex=SPY["2012-02-13/2012-02-15"])
}

