% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_conv}
\alias{roll_conv}
\title{Calculate the rolling convolutions (weighted sums) of a \emph{time series}
with a \emph{vector} of weights.}
\usage{
roll_conv(se_ries, weight_s)
}
\arguments{
\item{\code{se_ries}}{A \emph{time series} or a \emph{matrix} of data.}

\item{\code{weight_s}}{A column \emph{vector} of weights.}
}
\value{
A \emph{matrix} with the same dimensions as the input
  argument \code{se_ries}.
}
\description{
Calculate the rolling convolutions (weighted sums) of a \emph{time series}
with a \emph{vector} of weights.
}
\details{
The function \code{roll_conv()} calculates the convolutions of the
  \emph{matrix} columns with a \emph{vector} of weights.  It performs a loop
  down over the \emph{matrix} rows and multiplies the past (higher) values
  by the weights.  It calculates the rolling weighted sums of the past
  values.
  
  The function \code{roll_conv()} uses the \code{RcppArmadillo} function
  \code{arma::conv2()}. It performs a similar calculation to the standard
  \code{R} function \code{filter(x=se_ries, filter=weight_s,
  method="convolution", sides=1)}, but it's over \code{6} times faster, and
  it doesn't produce any leading \code{NA} values.
}
