% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{roll_moment}
\alias{roll_moment}
\title{Calculate a vector of statistics over an \code{OHLC} time series, and
calculate a rolling mean over the statistics.}
\usage{
roll_moment(oh_lc, mo_ment = "run_variance", win_dow = 11,
  weight_ed = TRUE, ...)
}
\arguments{
\item{oh_lc}{\code{OHLC} time series of prices and trading volumes, in
\code{xts} format.}

\item{mo_ment}{\code{character} string representing a function for 
estimating statistics of a single bar of \code{OHLC} data, such as
volatility, skew, and higher moments.}

\item{win_dow}{the size of the lookback window, equal to the number of bars
of data used for calculating the rolling mean.}

\item{weight_ed}{\code{Boolean} should statistic be weighted by trade volume?
(default \code{TRUE})}

\item{...}{additional parameters to the mo_ment function.}
}
\value{
An \code{xts} time series with a single column and the same number of
  rows as the argument \code{oh_lc}.
}
\description{
Calculate a vector of statistics over an \code{OHLC} time series, and
calculate a rolling mean over the statistics.
}
\details{
Calculates a vector of statistics over an \code{OHLC} time series, 
  such as volatility, skew, and higher moments.  The statistics could also be
  any other aggregation of a single bar of \code{OHLC} data, for example the
  \code{High} price minus the \code{Low} price squared.  The length of the 
  vector of statistics is equal to the number of rows of the argument 
  \code{oh_lc}. Then it calculates a trade volume weighted rolling mean over
  the vector of statistics over and calculate statistics.
}
\examples{
# create time index of one minute intervals over several days
in_dex <- seq(from=as.POSIXct("2015-02-09 09:30:00"),
              to=as.POSIXct("2015-02-28 16:00:00"), by="1 min")
# create synthetic xts of random prices
x_ts <- xts(exp(cumsum(0.001*rnorm(length(in_dex)))), order.by=in_dex)
# add trade Volume data
x_ts <- merge(x_ts,
          volume=sample(x=10*(2:18),
            size=length(in_dex), replace=TRUE))
# aggregate to hours OHLC data
oh_lc <- to.period(x=x_ts, period="hours")
# calculate time series of rolling variance and skew estimates
var_rolling <- roll_moment(oh_lc=oh_lc)
skew_rolling <- roll_moment(oh_lc=oh_lc, mo_ment="run_skew")
skew_rolling <- skew_rolling/(var_rolling)^(1.5)
skew_rolling[1, ] <- 0
skew_rolling <- na.locf(skew_rolling)
}

