% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{roll_sum}
\alias{roll_sum}
\title{Calculate the rolling sum of an \code{xts} time series over a sliding window
(lookback period).}
\usage{
roll_sum(x_ts, win_dow)
}
\arguments{
\item{x_ts}{an \code{xts} time series containing one or more columns of data.}

\item{win_dow}{an integer specifying the number of lookback periods.}
}
\value{
\code{xts} time series with the same dimensions as the input series.
}
\description{
Performs the same operation as function \code{runSum()} from package
\href{https://cran.r-project.org/web/packages/TTR/index.html}{TTR},
but using vectorized functions, so it's a little faster.
}
\details{
For example, if win_dow=3, then the rolling sum at any point is
  equal to the sum of \code{x_ts} values for that point plus two preceding
  points.
  The initial values of roll_sum() are equal to cumsum() values, so that
  roll_sum() doesn't return any NA values.
}
\examples{
# create xts time series
x_ts <- xts(x=rnorm(1000), order.by=(Sys.time()-3600*(1:1000)))
roll_sum(x_ts=get("SPY"), win_dow=3)
}

