% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_sum}
\alias{roll_sum}
\title{Calculate the rolling sum over a vector using \emph{Rcpp}.}
\usage{
roll_sum(vec_tor, look_back)
}
\arguments{
\item{vec_tor}{A numeric \emph{vector} of data.}

\item{look_back}{The length of the look-back interval, equal to the number 
of elements of data used for calculating the sum.}
}
\value{
A numeric \emph{vector} of the same length as the argument
  \code{vec_tor}.
}
\description{
Calculate the rolling sum over a vector using \emph{Rcpp}.
}
\details{
The function \code{roll_sum()} calculates a \emph{vector} of 
  rolling sums, over a \emph{vector} of data, using \emph{Rcpp}.  The
  function \code{roll_sum()} is over six times faster than
  \code{rutils::roll_sum()} which uses vectorized \emph{R}.
}
\examples{
\dontrun{
# calculate rolling sums over 11-period intervals
sum_rolling <- HighFreq::roll_sum(rnorm(1000), look_back=11)
}
}
