% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_sum}
\alias{roll_sum}
\title{Calculate the rolling sum over a \emph{vector} or a single-column \emph{time
series} using \emph{Rcpp}.}
\usage{
roll_sum(re_turns, look_back)
}
\arguments{
\item{\code{re_turns}}{A \emph{vector} or a single-column \emph{time series}.}

\item{\code{look_back}}{The length of the look-back interval, equal to the
number of elements of data used for calculating the sum.}
}
\value{
A column \emph{vector} of the same length as the argument
  \code{re_turns}.
}
\description{
Calculate the rolling sum over a \emph{vector} or a single-column \emph{time
series} using \emph{Rcpp}.
}
\details{
The function \code{roll_sum()} calculates a \emph{vector} of 
  rolling sums, over a \emph{vector} of data, using \emph{Rcpp}.  The
  function \code{roll_sum()} is several times faster than
  \code{rutils::roll_sum()} which uses vectorized \code{R} code.
}
\examples{
\dontrun{
# Create a vector of random returns
re_turns <- rnorm(1e6)
# Calculate rolling sums over 11-period lookback intervals
sum_rolling <- HighFreq::roll_sum(re_turns, look_back=11)
# Compare HighFreq::roll_sum() with rutils::roll_sum()
all.equal(HighFreq::roll_sum(re_turns, look_back=11), 
         rutils::roll_sum(re_turns, look_back=11))
# Compare the speed of Rcpp with R code
library(microbenchmark)
summary(microbenchmark(
  rcpp=HighFreq::roll_sum(re_turns, look_back=11),
  rcode=rutils::roll_sum(re_turns, look_back=11),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
