% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_sumep}
\alias{roll_sumep}
\title{Calculate the rolling sums at the end points of a \emph{time series} or a
\emph{matrix}.}
\usage{
roll_sumep(tseries, startp = 0L, endd = 0L, step = 1L, lookb = 1L, stub = 0L)
}
\arguments{
\item{\code{tseries}}{A \emph{time series} or a \emph{matrix}.}

\item{\code{startp}}{An \emph{integer} vector of start points (the default
is \code{startp = 0}).}

\item{\code{endd}}{An \emph{integer} vector of end points (the default is
\code{endd = 0}).}

\item{\code{step}}{The number of time periods between the end points (the
default is \code{step = 1}).}

\item{\code{lookb}}{The number of end points in the look-back interval
(the default is \code{lookb = 1}).}

\item{\code{stub}}{An \emph{integer} value equal to the first end point for
calculating the end points.}
}
\value{
A \emph{matrix} with the same number of columns as the input time
  series \code{tseries}, and the number of rows equal to the number of end
  points.
}
\description{
Calculate the rolling sums at the end points of a \emph{time series} or a
\emph{matrix}.
}
\details{
The function \code{roll_sumep()} calculates the rolling sums at the end
  points of the \emph{time series} \code{tseries}.
  
  The function \code{roll_sumep()} is implemented in \code{RcppArmadillo}
  \code{C++} code, which makes it several times faster than \code{R} code.
}
\examples{
\dontrun{
# Calculate historical returns
retp <- na.omit(rutils::etfenv$returns[, c("VTI", "IEF")])
# Define end points at 25 day intervals
endd <- HighFreq::calc_endpoints(NROW(retp), step=25)
# Define start points as 75 day lag of end points
startp <- HighFreq::calc_startpoints(endd, lookb=3)
# Calculate rolling sums using Rcpp
sumc <- HighFreq::roll_sumep(retp, startp=startp, endd=endd)
# Calculate rolling sums using R code
sumr <- sapply(1:NROW(endd), function(ep) {
colSums(retp[(startp[ep]+1):(endd[ep]+1), ])
  })  # end sapply
sumr <- t(sumr)
all.equal(sumc, sumr, check.attributes=FALSE)
}

}
