% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_var}
\alias{roll_var}
\title{Calculate a \emph{matrix} of dispersion (variance) estimates over a rolling
look-back interval attached at the end points of a \emph{time series} or a
\emph{matrix}.}
\usage{
roll_var(
  tseries,
  startp = 0L,
  endp = 0L,
  step = 1L,
  look_back = 1L,
  stub = 0L,
  method = "moment",
  con_fi = 0.75
)
}
\arguments{
\item{\code{tseries}}{A \emph{time series} or a \emph{matrix} of data.}

\item{\code{startp}}{An \emph{integer} vector of start points (the default
is \code{startp = 0}).}

\item{\code{endp}}{An \emph{integer} vector of end points (the default is
\code{endp = 0}).}

\item{\code{step}}{The number of time periods between the end points (the
default is \code{step = 1}).}

\item{\code{look_back}}{The number of end points in the look-back interval
(the default is \code{look_back = 1}).}

\item{\code{stub}}{An \emph{integer} value equal to the first end point for
calculating the end points (the default is \code{stub = 0}).}

\item{\code{method}}{A \emph{character} string representing the type of the
measure of dispersion (the default is \code{method = "moment"}).}
}
\value{
A \emph{matrix} with the same number of columns as the input time
  series \code{tseries}, and the number of rows equal to the number of end
  points.
}
\description{
Calculate a \emph{matrix} of dispersion (variance) estimates over a rolling
look-back interval attached at the end points of a \emph{time series} or a
\emph{matrix}.
}
\details{
The function \code{roll_var()} calculates a \emph{matrix} of dispersion
  (variance) estimates over rolling look-back intervals attached at the end
  points of the \emph{time series} \code{tseries}.
  
  The function \code{roll_var()} performs a loop over the end points, and at
  each end point it subsets the time series \code{tseries} over a look-back
  interval equal to \code{look_back} number of end points.
  
  It passes the subset time series to the function \code{calc_var()}, which
  calculates the dispersion.
  See the function \code{calc_var()} for a description of the dispersion
  methods.
  
  If the arguments \code{endp} and \code{startp} are not given then it
  first calculates a vector of end points separated by \code{step} time
  periods. It calculates the end points along the rows of \code{tseries}
  using the function \code{calc_endpoints()}, with the number of time
  periods between the end points equal to \code{step} time periods.

  For example, the rolling variance at \code{25} day end points, with a
  \code{75} day look-back, can be calculated using the parameters
  \code{step = 25} and \code{look_back = 3}.

  The function \code{roll_var()} with the parameter \code{step = 1}
  performs the same calculation as the function \code{roll_var()} from
  package
  \href{https://cran.r-project.org/web/packages/RcppRoll/index.html}{RcppRoll},
  but it's several times faster because it uses \code{RcppArmadillo}
  \code{C++} code.

  The function \code{roll_var()} is implemented in \code{RcppArmadillo}
  \code{C++} code, so it's many times faster than the equivalent \code{R}
  code.
}
\examples{
\dontrun{
# Define time series of returns using package rutils
re_turns <- na.omit(rutils::etf_env$re_turns$VTI)
# Calculate the rolling variance at 25 day end points, with a 75 day look-back
vari_ance <- HighFreq::roll_var(re_turns, step=25, look_back=3)
# Compare the variance estimates over 11-period lookback intervals
all.equal(HighFreq::roll_var(re_turns, look_back=11)[-(1:10), ], 
  drop(RcppRoll::roll_var(re_turns, n=11)), check.attributes=FALSE)
# Compare the speed of HighFreq::roll_var() with RcppRoll::roll_var()
library(microbenchmark)
summary(microbenchmark(
  Rcpp=HighFreq::roll_var(re_turns, look_back=11),
  RcppRoll=RcppRoll::roll_var(re_turns, n=11),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
# Compare the speed of HighFreq::roll_var() with TTR::runMAD()
summary(microbenchmark(
    Rcpp=HighFreq::roll_var(re_turns, look_back=11, method="quantile"),
    TTR=TTR::runMAD(re_turns, n = 11),
    times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}
}
