% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_var_vec}
\alias{roll_var_vec}
\title{Calculate a \emph{vector} of variance estimates over a rolling look-back
interval for a single-column \emph{time series} or a \emph{vector}, using
\code{RcppArmadillo}.}
\usage{
roll_var_vec(se_ries, look_back = 11L)
}
\arguments{
\item{\code{se_ries}}{A single-column \emph{time series} or a \emph{vector}.}

\item{\code{look_back}}{The length of the look-back interval, equal to the
number of \emph{vector} elements used for calculating a single variance
estimate.}
}
\value{
A column \emph{vector} with the same number of elements as the input
  argument \code{se_ries}.
}
\description{
Calculate a \emph{vector} of variance estimates over a rolling look-back
interval for a single-column \emph{time series} or a \emph{vector}, using
\code{RcppArmadillo}.
}
\details{
The function \code{roll_var_vec()} calculates a \emph{vector} of
  variance estimates over a rolling look-back interval for a single-column
  \emph{time series} or a \emph{vector}, using \code{RcppArmadillo}
  \code{C++} code.
  
  The function \code{roll_var_vec()} uses an expanding look-back interval in
  the initial warmup period, to calculate the same number of elements as the
  input argument \code{se_ries}.

  The function \code{roll_var_vec()} performs the same calculation as the
  function \code{roll_var()} from package
  \href{https://cran.r-project.org/web/packages/RcppRoll/index.html}{RcppRoll},
  but it's several times faster because it uses \code{RcppArmadillo}
  \code{C++} code.
}
\examples{
\dontrun{
# Create a vector of random returns
re_turns <- rnorm(1e6)
# Compare the variance estimates over 11-period lookback intervals
all.equal(drop(HighFreq::roll_var_vec(re_turns, look_back=11))[-(1:10)], 
  RcppRoll::roll_var(re_turns, n=11))
# Compare the speed of RcppArmadillo with RcppRoll
library(microbenchmark)
summary(microbenchmark(
  RcppArmadillo=HighFreq::roll_var_vec(re_turns, look_back=11),
  RcppRoll=RcppRoll::roll_var(re_turns, n=11),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}
}
