% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{roll_vwap}
\alias{roll_vwap}
\title{Calculate the volume-weighted average price of an \code{OHLC} time series
over a rolling window (lookback period).}
\usage{
roll_vwap(oh_lc, x_ts = oh_lc[, 4], win_dow)
}
\arguments{
\item{oh_lc}{an \code{OHLC} time series of prices in \code{xts} format.}

\item{x_ts}{single-column \code{xts} time series.}

\item{win_dow}{the size of the lookback window, equal to the number of bars
of data used for calculating the average price.}
}
\value{
An \code{xts} time series with a single column and the same number of
  rows as the argument \code{oh_lc}.
}
\description{
Performs the same operation as function \code{VWAP()} from package
\href{https://cran.r-project.org/web/packages/TTR/index.html}{VWAP},
but using vectorized functions, so it's a little faster.
}
\details{
The function \code{roll_vwap()} calculates the volume-weighted
  average closing price, defined as the sum of the prices multiplied by
  trading volumes in the lookback window, divided by the sum of trading
  volumes in the window. If the argument \code{x_ts} is passed in explicitly,
  then its volume-weighted average value over time is calculated.
}
\examples{
# calculate and plot rolling volume-weighted average closing prices (VWAP)
prices_rolling <- roll_vwap(oh_lc=SPY["2013-11"], win_dow=11)
chart_Series(SPY["2013-11-12"], name="SPY prices")
add_TA(prices_rolling["2013-11-12"], on=1, col="red", lwd=2)
legend("top", legend=c("SPY prices", "VWAP prices"),
bg="white", lty=c(1, 1), lwd=c(2, 2),
col=c("black", "red"), bty="n")
# calculate running returns
returns_running <- run_returns(x_ts=SPY)
# calculate the rolling volume-weighted average returns
roll_vwap(oh_lc=SPY, x_ts=returns_running, win_dow=11)
}

