% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_zscores}
\alias{roll_zscores}
\title{Calculate a \emph{vector} of z-scores of the residuals of rolling
regressions at the end points of the predictor matrix.}
\usage{
roll_zscores(
  respv,
  predm,
  startp = 0L,
  endd = 0L,
  step = 1L,
  lookb = 1L,
  stub = 0L
)
}
\arguments{
\item{\code{respv}}{A single-column \emph{time series} or a \emph{vector}
of response data.}

\item{\code{predm}}{A \emph{time series} or a \emph{matrix} of predictor
data.}

\item{\code{startp}}{An \emph{integer} vector of start points (the default
is \code{startp = 0}).}

\item{\code{endd}}{An \emph{integer} vector of end points (the default is
\code{endd = 0}).}

\item{\code{step}}{The number of time periods between the end points (the
default is \code{step = 1}).}

\item{\code{lookb}}{The number of end points in the look-back interval
(the default is \code{lookb = 1}).}

\item{\code{stub}}{An \emph{integer} value equal to the first end point for
calculating the end points (the default is \code{stub = 0}).}
}
\value{
A column \emph{vector} of the same length as the number of rows of
  \code{predm}.
}
\description{
Calculate a \emph{vector} of z-scores of the residuals of rolling
regressions at the end points of the predictor matrix.
}
\details{
The function \code{roll_zscores()} calculates a \emph{vector} of z-scores
  of the residuals of rolling regressions at the end points of the
  \emph{time series} \code{predm}.
  
  The function \code{roll_zscores()} performs a loop over the end points,
  and at each end point it subsets the time series \code{predm} over a
  look-back interval equal to \code{lookb} number of end points.
  
  It passes the subset time series to the function \code{calc_lm()}, which
  calculates the regression data.
  
  If the arguments \code{endd} and \code{startp} are not given then it
  first calculates a vector of end points separated by \code{step} time
  periods. It calculates the end points along the rows of \code{predm}
  using the function \code{calc_endpoints()}, with the number of time
  periods between the end points equal to \code{step} time periods.
  
  For example, the rolling variance at \code{25} day end points, with a
  \code{75} day look-back, can be calculated using the parameters
  \code{step = 25} and \code{lookb = 3}.
}
\examples{
\dontrun{
# Calculate historical returns
retp <- na.omit(rutils::etfenv$returns[, c("XLF", "VTI", "IEF")])
# Response equals XLF returns
respv <- retp[, 1]
# Predictor matrix equals VTI and IEF returns
predm <- retp[, -1]
# Calculate Z-scores from rolling time series regression using RcppArmadillo
lookb <- 11
zscores <- HighFreq::roll_zscores(respv=respv, predm=predm, lookb)
# Calculate z-scores in R from rolling multivariate regression using lm()
zscoresr <- sapply(1:NROW(predm), function(ro_w) {
  if (ro_w == 1) return(0)
  startpoint <- max(1, ro_w-lookb+1)
  responsi <- response[startpoint:ro_w]
  predicti <- predictor[startpoint:ro_w, ]
  regmod <- lm(responsi ~ predicti)
  residuals <- regmod$residuals
  residuals[NROW(residuals)]/sd(residuals)
})  # end sapply
# Compare the outputs of both functions
all.equal(zscores[-(1:lookb)], zscoresr[-(1:lookb)], 
  check.attributes=FALSE)
}

}
