% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{run_skew}
\alias{run_skew}
\title{Calculate time series of skew estimates from a \code{OHLC} time series.}
\usage{
run_skew(oh_lc, calc_method = "rogers.satchell")
}
\arguments{
\item{oh_lc}{an \code{OHLC} time series of prices in \code{xts} format.}

\item{calc_method}{\code{character} string representing method for estimating
skew.}
}
\value{
Time series of skew estimates.
}
\description{
Calculate time series of skew estimates from a \code{OHLC} time series.
}
\details{
Calculates skew estimates from \code{OHLC} prices at each
  point in time (row).  The methods include Garman-Klass and Rogers-Satchell.
}
\examples{
# create time index of one second intervals for a single day
in_dex <- seq(from=as.POSIXct("2015-02-09 09:30:00"),
              to=as.POSIXct("2015-02-09 16:00:00"), by="1 sec")
# create synthetic xts of random prices
x_ts <- xts(cumsum(rnorm(length(in_dex))), order.by=in_dex)
# aggregate to minutes OHLC data
oh_lc <- to.period(x=x_ts, period="minutes")
# calculate skew estimates
sk_ew <- run_skew(oh_lc)
}

