% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{run_variance}
\alias{run_variance}
\title{Calculate a time series of variance estimates for an \code{OHLC} time series.}
\usage{
run_variance(oh_lc, calc_method = "garman.klass_yz")
}
\arguments{
\item{oh_lc}{an \code{OHLC} time series of prices in \code{xts} format.}

\item{calc_method}{\code{character} string representing method for estimating
  variance.  The methods include:
\itemize{
  \item "close" close to close,
  \item "garman.klass" Garman-Klass,
  \item "garman.klass_yz" Garman-Klass with account for close-to-open price jumps,
  \item "rogers.satchell" Rogers-Satchell,
  \item "yang.zhang" Yang-Zhang,
}}
}
\value{
An \code{xts} time series with a single column and the same number of
  rows as the argument \code{oh_lc}.
}
\description{
Calculates the variance estimates for each bar of \code{OHLC} prices at each
point in time (row), using the squared differences of \code{OHLC} prices at
each point in time.
}
\details{
The function \code{run_variance()} performs a similar operation to
  function \code{volatility()} from package
  \href{https://cran.r-project.org/web/packages/TTR/index.html}{TTR}, but
  without calculating a running sum using \code{runSum()}.  It's also a 
  little faster because it performs less data validation. The variance 
  estimation methods "close", "garman.klass_yz", and "yang.zhang" do account 
  for close-to-open price jumps, while the methods "garman.klass" and 
  "rogers.satchell" do not account for close-to-open price jumps.
}
\examples{
# create minutely OHLC time series of random prices
oh_lc <- HighFreq::random_OHLC()
# calculate variance estimates for oh_lc
var_running <- HighFreq::run_variance(oh_lc)
# calculate variance estimates for SPY
var_running <- HighFreq::run_variance(SPY, calc_method="yang.zhang")
# calculate SPY variance without overnight jumps
var_running <- HighFreq::run_variance(SPY, calc_method="rogers.satchell")
}

