% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{save_rets}
\alias{save_rets}
\title{Load, scrub, aggregate, and rbind multiple days of \emph{TAQ} data for a
single symbol. Calculate returns and save them to a single
\sQuote{\code{*.RData}} file.}
\usage{
save_rets(sym_bol, data_dir = "E:/mktdata/sec/",
  output_dir = "E:/output/data/", look_back = 51, vol_mult = 2,
  period = "minutes", tzone = "America/New_York")
}
\value{
A time series of returns and volume in \emph{xts} format.
}
\description{
Load, scrub, aggregate, and rbind multiple days of \emph{TAQ} data for a
single symbol. Calculate returns and save them to a single
\sQuote{\code{*.RData}} file.
}
\details{
The function \code{save_rets} loads multiple days of \emph{TAQ}
  data, then scrubs, aggregates, and rbinds them into a \emph{OHLC} time
  series.  It then calculates returns using function \code{run_returns()}, and
  stores them in a variable named \sQuote{\code{symbol.rets}}, and saves them
  to a file called \sQuote{\code{symbol.rets.RData}}.
  The \emph{TAQ} data files are assumed to be stored in separate directories
  for each \sQuote{\code{symbol}}. Each \sQuote{\code{symbol}} has its own
  directory (named \sQuote{\code{symbol}}) in the \sQuote{\code{data_dir}}
  directory. Each \sQuote{\code{symbol}} directory contains multiple daily
  \sQuote{\code{*.RData}} files, each file containing one day of \emph{TAQ}
  data.
}
\examples{
\dontrun{
save_rets("SPY")
}
}
