% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{save_scrub_agg}
\alias{save_scrub_agg}
\title{Load, scrub, aggregate, and rbind multiple days of \emph{TAQ} data for a
single symbol, and save the \emph{OHLC} time series to a single
\sQuote{\code{*.RData}} file.}
\usage{
save_scrub_agg(
  symbol,
  dirin = "E:/mktdata/sec/",
  dirout = "E:/output/data/",
  lookb = 51,
  volm = 2,
  period = "minutes",
  tzone = "America/New_York"
)
}
\arguments{
\item{\code{symbol}}{A \emph{character} string representing symbol or ticker.}

\item{\code{dirin}}{A \emph{character} string representing directory
containing input \sQuote{\code{*.RData}} files.}

\item{\code{dirout}}{A \emph{character} string representing directory
containing output \sQuote{\code{*.RData}} files.}
}
\value{
An \emph{OHLC} time series in \emph{xts} format.
}
\description{
Load, scrub, aggregate, and rbind multiple days of \emph{TAQ} data for a
single symbol, and save the \emph{OHLC} time series to a single
\sQuote{\code{*.RData}} file.
}
\details{
The function \code{save_scrub_agg()} loads multiple days of
  \emph{TAQ} data, then scrubs, aggregates, and rbinds them into a
  \emph{OHLC} time series, and finally saves it to a single
  \sQuote{\code{*.RData}} file. The \emph{OHLC} time series is stored in a
  variable named \sQuote{\code{symbol}}, and then it's saved to a file named
  \sQuote{\code{symbol.RData}} in the \sQuote{\code{dirout}} directory.
  The \emph{TAQ} data files are assumed to be stored in separate directories
  for each \sQuote{\code{symbol}}. Each \sQuote{\code{symbol}} has its own
  directory (named \sQuote{\code{symbol}}) in the \sQuote{\code{dirin}}
  directory. Each \sQuote{\code{symbol}} directory contains multiple daily
  \sQuote{\code{*.RData}} files, each file containing one day of \emph{TAQ}
  data.
}
\examples{
\dontrun{
# set data directories
dirin <- "C:/Develop/data/hfreq/src/"
dirout <- "C:/Develop/data/hfreq/scrub/"
symbol <- "SPY"
# Aggregate SPY TAQ data to 15-min OHLC bar data, and save the data to a file
save_scrub_agg(symbol=symbol, dirin=dirin, dirout=dirout, period="15 min")
}

}
