% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{save_scrub_agg}
\alias{save_scrub_agg}
\title{Load, scrub, aggregate, and rbind multiple days of \emph{TAQ} data for a
single symbol, and save the \emph{OHLC} time series to a single
\sQuote{\code{*.RData}} file.}
\usage{
save_scrub_agg(sym_bol, data_dir = "E:/mktdata/sec/",
  output_dir = "E:/output/data/", look_back = 51, vol_mult = 2,
  period = "minutes", tzone = "America/New_York")
}
\arguments{
\item{sym_bol}{A \emph{character} string representing symbol or ticker.}

\item{data_dir}{A \emph{character} string representing directory containing
input \sQuote{\code{*.RData}} files.}

\item{output_dir}{A \emph{character} string representing directory containing
output \sQuote{\code{*.RData}} files.}

\item{look_back}{The number of data points in rolling look-back interval for
estimating rolling quantile.}

\item{vol_mult}{The quantile multiplier.}

\item{period}{The aggregation period.}

\item{tzone}{The timezone to convert.}
}
\value{
An \emph{OHLC} time series in \emph{xts} format.
}
\description{
Load, scrub, aggregate, and rbind multiple days of \emph{TAQ} data for a
single symbol, and save the \emph{OHLC} time series to a single
\sQuote{\code{*.RData}} file.
}
\details{
The function \code{save_scrub_agg()} loads multiple days of
  \emph{TAQ} data, then scrubs, aggregates, and rbinds them into a
  \emph{OHLC} time series, and finally saves it to a single
  \sQuote{\code{*.RData}} file. The \emph{OHLC} time series is stored in a
  variable named \sQuote{\code{symbol}}, and then it's saved to a file named
  \sQuote{\code{symbol.RData}} in the \sQuote{\code{output_dir}} directory.
  The \emph{TAQ} data files are assumed to be stored in separate directories
  for each \sQuote{\code{symbol}}. Each \sQuote{\code{symbol}} has its own
  directory (named \sQuote{\code{symbol}}) in the \sQuote{\code{data_dir}}
  directory. Each \sQuote{\code{symbol}} directory contains multiple daily
  \sQuote{\code{*.RData}} files, each file containing one day of \emph{TAQ}
  data.
}
\examples{
\dontrun{
# set data directories
data_dir <- "C:/Develop/data/hfreq/src/"
output_dir <- "C:/Develop/data/hfreq/scrub/"
sym_bol <- "SPY"
# aggregate SPY TAQ data to 15-min OHLC bar data, and save the data to a file
save_scrub_agg(sym_bol=sym_bol, data_dir=data_dir, output_dir=output_dir, period="15 min")
}
}

