% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{save_taq}
\alias{save_taq}
\title{Load and scrub multiple days of \emph{TAQ} data for a single symbol, and save
it to multiple \sQuote{\code{*.RData}} files.}
\usage{
save_taq(sym_bol, data_dir = "E:/mktdata/sec/",
  output_dir = "E:/output/data/", look_back = 51, vol_mult = 2,
  tzone = "America/New_York")
}
\arguments{
\item{sym_bol}{A \emph{character} string representing symbol or ticker.}

\item{data_dir}{A \emph{character} string representing directory containing
input \sQuote{\code{*.RData}} files.}

\item{output_dir}{A \emph{character} string representing directory containing
output \sQuote{\code{*.RData}} files.}

\item{look_back}{The number of data points in rolling look-back interval for 
estimating rolling quantile.}

\item{vol_mult}{The quantile multiplier.}

\item{tzone}{The timezone to convert.}
}
\value{
a \emph{TAQ} time series in \emph{xts} format.
}
\description{
Load and scrub multiple days of \emph{TAQ} data for a single symbol, and save
it to multiple \sQuote{\code{*.RData}} files.
}
\details{
The function \code{save_taq()} loads multiple days of \emph{TAQ}
  data, scrubs it, and saves the scrubbed TAQ data to individual
  \sQuote{\code{*.RData}} files. It uses the same file names for output as
  the input file names. The \emph{TAQ} data files are assumed to be stored in
  separate directories for each \sQuote{\code{symbol}}. Each
  \sQuote{\code{symbol}} has its own directory (named \sQuote{\code{symbol}})
  in the \sQuote{\code{data_dir}} directory.
  Each \sQuote{\code{symbol}} directory contains multiple daily
  \sQuote{\code{*.RData}} files, each file containing one day of \emph{TAQ}
  data.
}
\examples{
\dontrun{
save_taq("SPY")
}
}
