% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{save_TAQ}
\alias{save_TAQ}
\title{Load and scrub multiple days of \code{TAQ} data for a single symbol, and save
it to multiple \sQuote{\code{*.RData}} files.}
\usage{
save_TAQ(sym_bol, data_dir = "E:/mktdata/sec/",
  output_dir = "E:/output/data/", win_dow = 51, vol_mult = 2,
  tzone = "America/New_York")
}
\arguments{
\item{sym_bol}{\code{character} string representing symbol or ticker.}

\item{data_dir}{\code{character} string representing directory containing
input \sQuote{\code{*.RData}} files.}

\item{output_dir}{\code{character} string representing directory containing
output \sQuote{\code{*.RData}} files.}

\item{win_dow}{number of data points for estimating rolling volatility.}

\item{vol_mult}{volatility multiplier.}

\item{tzone}{timezone to convert.}
}
\value{
\code{TAQ} time series in \code{xts} format.
}
\description{
Load and scrub multiple days of \code{TAQ} data for a single symbol, and save
it to multiple \sQuote{\code{*.RData}} files.
}
\details{
The function \code{save_TAQ} loads multiple days of \code{TAQ} data,
  scrubs it, and saves it to \sQuote{\code{*.RData}} files. It uses the same
  file names for output as the input file names.
  The \code{TAQ} data files are assumed to be stored in separate directories
  for each \sQuote{\code{symbol}}. Each \sQuote{\code{symbol}} has its own
  directory (named \sQuote{\code{symbol}}) in the \sQuote{\code{data_dir}}
  directory. Each \sQuote{\code{symbol}} directory contains multiple daily
  \sQuote{\code{*.RData}} files, each file containing one day of \code{TAQ} data.
}
\examples{
\dontrun{
save_TAQ("SPY")
}
}

