% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{scrub_agg}
\alias{scrub_agg}
\title{Scrub a single day of \code{TAQ} data, aggregate it, and convert to
\code{OHLC} format.}
\usage{
scrub_agg(ta_q, win_dow = 51, vol_mult = 2, period = "minutes",
  tzone = "America/New_York")
}
\arguments{
\item{ta_q}{\code{TAQ} time series in \code{xts} format.}

\item{win_dow}{number of data points for estimating rolling volatility.}

\item{vol_mult}{volatility multiplier.}

\item{period}{aggregation period.}

\item{tzone}{timezone to convert.}
}
\value{
A \code{OHLC} time series in \code{xts} format.
}
\description{
Scrub a single day of \code{TAQ} data, aggregate it, and convert to
\code{OHLC} format.
}
\details{
The function \code{scrub_agg()} performs:
\itemize{
  \item index timezone conversion,
  \item data subset to trading hours,
  \item removal of duplicate time stamps,
  \item scrubbing of quotes with suspect bid-offer spreads,
  \item scrubbing of quotes with suspect price jumps,
  \item cbinding of mid prices with volume data,
  \item aggregation to OHLC using function \code{to.period} from package \code{xts},
}
Valid 'period' character strings include: "minutes", "3 min", "5 min", "10
min", "15 min", "30 min", and "hours". The time index of the output time
series is rounded up to the next integer multiple of 'period'.
}
\examples{
# create random TAQ prices
ta_q <- HighFreq::random_taq()
# aggregate to ten minutes OHLC data
oh_lc <- HighFreq::scrub_agg(ta_q, period="10 min")
chart_Series(oh_lc, name="random prices")
# scrub and aggregate a single day of SPY TAQ data to OHLC
oh_lc <- HighFreq::scrub_agg(ta_q=SPY_TAQ)
chart_Series(oh_lc, name=sym_bol)
}

