% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{scrub_agg}
\alias{scrub_agg}
\title{Scrub a single day of \emph{TAQ} data, aggregate it, and convert to
\emph{OHLC} format.}
\usage{
scrub_agg(
  taq,
  lookb = 51,
  volm = 2,
  period = "minutes",
  tzone = "America/New_York"
)
}
\arguments{
\item{\code{period}}{The aggregation period.}
}
\value{
A \emph{OHLC} time series in \emph{xts} format.
}
\description{
Scrub a single day of \emph{TAQ} data, aggregate it, and convert to
\emph{OHLC} format.
}
\details{
The function \code{scrub_agg()} performs:
\itemize{
  \item index timezone conversion,
  \item data subset to trading hours,
  \item removal of duplicate time stamps,
  \item scrubbing of quotes with suspect bid-ask spreads,
  \item scrubbing of quotes with suspect price jumps,
  \item cbinding of mid prices with volume data,
  \item aggregation to OHLC using function \code{to.period()} from package \emph{xts},
}
Valid 'period' character strings include: "minutes", "3 min", "5 min", "10
min", "15 min", "30 min", and "hours". The time index of the output time
series is rounded up to the next integer multiple of 'period'.
}
\examples{
# Create random TAQ prices
taq <- HighFreq::random_taq()
# Aggregate to ten minutes OHLC data
ohlc <- HighFreq::scrub_agg(taq, period="10 min")
chart_Series(ohlc, name="random prices")
# scrub and aggregate a single day of SPY TAQ data to OHLC
ohlc <- HighFreq::scrub_agg(taq=HighFreq::SPY_TAQ)
chart_Series(ohlc, name=symbol)

}
