% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{scrub_TAQ}
\alias{scrub_TAQ}
\title{Scrub a single day of TAQ data in xts format.}
\usage{
scrub_TAQ(taq_data, vol_window = 51, thresh_old = 2)
}
\arguments{
\item{taq_data}{\code{TAQ xts} time series.}

\item{vol_window}{number of data points for estimating running volatility.}

\item{thresh_old}{level of prices above which values are considered extreme.}
}
\value{
\code{TAQ xts} time series.
}
\description{
Scrub a single day of TAQ data in xts format.
}
\details{
Jump values are those where neighboring values both exceed the
  scaled running volatility (defined as a quantile), but their sum doesn't
  exceed it.
}
\examples{
# create time index of one second intervals for a single day
in_dex <- seq(from=as.POSIXct("2015-02-09 09:30:00"), to=as.POSIXct("2015-02-09 16:00:00"), by="1 sec")
# create xts of random prices
x_ts <- xts(cumsum(rnorm(length(in_dex))), order.by=in_dex)
# create vector of random bid-offer prices
bid_offer <- abs(rnorm(length(in_dex)))/10
# create TAQ data using cbind
taq_data <- cbind(x_ts-bid_offer, x_ts+bid_offer)
# add Trade.Price
taq_data <- cbind(taq_data, x_ts+rnorm(length(in_dex))/10)
# add Volume
taq_data <- cbind(taq_data, sample(x=10*(2:18), size=length(in_dex), replace=TRUE))
colnames(taq_data) <- c("Bid.Price", "Ask.Price", "Trade.Price", "Volume")
taq_data <- scrub_TAQ(taq_data)
taq_data <- scrub_TAQ(taq_data, vol_window=11, thresh_old=1)
}

