% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{season_ality}
\alias{season_ality}
\title{Perform daily, weekly, monthly, and yearly seasonality aggregations over a
univariate \code{xts} time series.}
\usage{
season_ality(x_ts)
}
\arguments{
\item{x_ts}{univariate \code{xts} time series.}
}
\value{
\code{xts} time series with aggregations over the seasonality
  interval.
}
\description{
Perform daily, weekly, monthly, and yearly seasonality aggregations over a
univariate \code{xts} time series.
}
\details{
An example of a daily seasonality aggregation is the average price
  of a stock between 9:30AM and 10:00AM every day, over many days.
}
\examples{
season_ality(x_ts=get("SPY"))
}

