% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighFreq.R
\name{season_ality}
\alias{season_ality}
\title{Perform seasonality aggregations over a single-column \emph{xts} time series.}
\usage{
season_ality(x_ts, in_dex = format(zoo::index(x_ts), "\%H:\%M"))
}
\arguments{
\item{x_ts}{A single-column \emph{xts} time series.}

\item{in_dex}{A vector of \emph{character} strings representing points in
time, of the same length as the argument \code{x_ts}.}
}
\value{
An \emph{xts} time series with mean aggregations over the seasonality
  interval.
}
\description{
Perform seasonality aggregations over a single-column \emph{xts} time series.
}
\details{
The function \code{season_ality()} calculates the mean of values
  observed at the same points in time specified by the argument
  \code{in_dex}. An example of a daily seasonality aggregation is the average
  price of a stock between 9:30AM and 10:00AM every day, over many days. The
  argument \code{in_dex} is passed into function \code{tapply()}, and must be
  the same length as the argument \code{x_ts}.
}
\examples{
# Calculate running variance of each minutely OHLC bar of data
x_ts <- run_variance(HighFreq::SPY)
# Remove overnight variance spikes at "09:31"
in_dex <- format(index(x_ts), "\%H:\%M")
x_ts <- x_ts[!in_dex=="09:31", ]
# Calculate daily seasonality of variance
var_seasonal <- season_ality(x_ts=x_ts)
chart_Series(x=var_seasonal, name=paste(colnames(var_seasonal),
  "daily seasonality of variance"))
}
