% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_garch}
\alias{sim_garch}
\title{Simulate a \emph{GARCH} process using \emph{Rcpp}.}
\usage{
sim_garch(om_ega, al_pha, be_ta, in_nov)
}
\arguments{
\item{\code{om_ega}}{Parameter proportional to the long-term average level
of variance.}

\item{\code{al_pha}}{The weight associated with recent realized variance
updates.}

\item{\code{be_ta}}{The weight associated with the past variance estimates.}

\item{\code{in_nov}}{A \emph{vector} of innovations (random numbers).}
}
\value{
A \emph{matrix} with two columns: the simulated returns and
  variance, and with the same number of rows as the length of the argument 
  \code{in_nov}.
}
\description{
Simulate a \emph{GARCH} process using \emph{Rcpp}.
}
\details{
The function \code{sim_garch()} simulates a \emph{GARCH} process
  using \emph{Rcpp}.
}
\examples{
\dontrun{
# Define the GARCH model parameters
om_ega <- 0.01
al_pha <- 0.5
be_ta <- 0.2
# Simulate the GARCH process using Rcpp
garch_rcpp <- sim_garch(om_ega=om_ega, al_pha=al_pha, be_ta=be_ta, in_nov=rnorm(10000))
}

}
