% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_ou}
\alias{sim_ou}
\title{Simulate an \emph{Ornstein-Uhlenbeck} process using \emph{Rcpp}.}
\usage{
sim_ou(eq_price, vol_at, the_ta, in_nov)
}
\arguments{
\item{eq_price}{The equilibrium price.}

\item{vol_at}{The volatility of returns.}

\item{the_ta}{The strength of mean reversion.}

\item{in_nov}{A \emph{vector} of innovations (random numbers).}
}
\value{
A \emph{vector} representing the \emph{time series} of prices, with
  the same length as the argument \code{in_nov}.
}
\description{
Simulate an \emph{Ornstein-Uhlenbeck} process using \emph{Rcpp}.
}
\details{
The function \code{sim_ou()} simulates an \emph{Ornstein-Uhlenbeck}
  process using \emph{Rcpp}, and returns a \emph{vector} representing the 
  \emph{time series} of prices.
}
\examples{
\dontrun{
# Define the Ornstein-Uhlenbeck model parameters
eq_price <- 5.0
vol_at <- 0.01
the_ta <- 0.01
# Simulate Ornstein-Uhlenbeck process using Rcpp
price_s <- HighFreq::sim_ou_rcpp(eq_price=eq_price, vol_at=vol_at, the_ta=the_ta, in_nov=rnorm(1000))
}

}
