% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vari_ance}
\alias{vari_ance}
\title{Calculate the variance of a vector using \emph{Rcpp}.}
\usage{
vari_ance(vec_tor)
}
\arguments{
\item{vec_tor}{A numeric \emph{vector} of data.}
}
\value{
A single \emph{numeric} value.
}
\description{
Calculate the variance of a vector using \emph{Rcpp}.
}
\details{
The function \code{vari_ance()} calculates the variance of a vector
  using \emph{Rcpp}. The function \code{vari_ance()} is slightly faster than
  the \emph{R} function \code{var()}.
}
\examples{
\dontrun{
# calculate variance 
HighFreq::vari_ance(rnorm(1000))
}
}
